% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{addChild}
\alias{addChild}
\title{Add child node to .self}
\arguments{
\item{isLeftChild}{Boolean, TRUE if child to add is left child, FALSE if child to add is right child.}
}
\description{
Add child node to .self
}
