% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{addItem}
\alias{addItem}
\title{Add numeric value itemtoAdd to node's weight.}
\arguments{
\item{itemToAdd}{Numeric value}

\item{targetNodeIndex}{Integer, target index of the node whose weight is to be added to.}
}
\description{
Add numeric value itemtoAdd to node's weight.
}
