% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{addItems}
\alias{addItems}
\title{Add values to the weights of nodes in Tree object.}
\arguments{
\item{itemsToAdd}{Array of numeric values.}
}
\description{
The nth item in itemsToAdd is added to the weight of the node with index n.
}
\examples{
t <- Tree$new(c(0,4),2) # this will be a tree with 2 leaves and 3 total nodes including the root
t$addItems(c(1,3,7)) # this will add weight 1 to the root, weight 3 to the node with index 2, and weight 7 to the node with index 3.
t$printTree() 
}
