% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{adjacentElements}
\alias{adjacentElements}
\title{Creates array of adjacent elements in almost certainly a not optimally-efficient way}
\usage{
adjacentElements(ls)
}
\arguments{
\item{ls}{1D array of even length n}
}
\value{
1D array of length n equal to ls with every other element swapped
}
\description{
Creates array of adjacent elements in almost certainly a not optimally-efficient way
}
\examples{
ls <- c(1,2,3,4)
adjacentElements(ls) #outputs c(2,1,3,4)

}
