% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{amsweep}
\alias{amsweep}
\title{Sweep operator}
\usage{
amsweep(g, m, reverse = FALSE)
}
\arguments{
\item{g}{Each unit of a numeric, symmetric covariance matrix divided by
the number of observations in the data the covariance matrix was 
derived from.}

\item{m}{A logical vector of length equal to the number of rows in \code{g}
in which the \code{TRUE} values correspond to the x values in the matrix
and the \code{FALSE} values correspond to the y value in the matrix.}

\item{reverse}{Logical vector specifying whether the sign of the matrix 
should be flipped. Default to \code{FALSE}.}
}
\value{
The coefficients from \code{g}.
}
\description{
General sweep operator citation:
}
\references{
Goodnight, James H. "A tutorial on the SWEEP operator." The American Statistician 33, no. 3 (1979): 149-158.
This implementation is from pseudocode from:

Schafer, Joseph L. Analysis of incomplete multivariate data. Chapman and Hall/CRC, 1997.
Code ported from:

Honaker, James, Gary King, and Matthew Blackwell. "Amelia II: A program for missing data." Journal of statistical software 45, no. 7 (2011): 1-47.

Sweeps a covariance matrix to extract regression coefficients.
}
