% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{binDataPoint}
\alias{binDataPoint}
\title{Helper function to bin a single data point into the tree.
Traverses the tree structure until the data point has been binned at all levels of the tree.}
\arguments{
\item{x}{Numeric. Data point to be binned}

\item{currNode}{Node object. The current node to bin point into.}
}
\description{
Helper function to bin a single data point into the tree.
Traverses the tree structure until the data point has been binned at all levels of the tree.
}
