% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mean.R
\name{boot.mean}
\alias{boot.mean}
\title{Mean function}
\usage{
boot.mean(xi, n)
}
\arguments{
\item{xi}{Vector of values}

\item{n}{Number of observations}
}
\value{
Mean
}
\description{
This function is used when the bootstrap mechanism is used
}
