% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{bootHist}
\alias{bootHist}
\title{Bootstrap replication for histogram}
\usage{
bootHist(xi, varType, bins = NULL)
}
\arguments{
\item{xi}{Bootstrapped vector of numeric or categorical type.}

\item{varType}{Character indicating the variable type.}

\item{bins}{Vector indicating the bins into which \code{xi} is partitioned.}
}
\value{
Histogram with counts for each level of \code{xi}.
}
\description{
This is a wrapper for the histogram function used internally by the
bootstrap mechanism.
}
