% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mean.R
\name{bootMean}
\alias{bootMean}
\title{Bootstrap mean function}
\usage{
bootMean(xi, n)
}
\arguments{
\item{xi}{Vector of values}

\item{n}{Number of observations}
}
\value{
Mean
}
\description{
This function is used when the bootstrap mechanism is used
}
