% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism-bootstrap.R
\name{bootstrap.replication}
\alias{bootstrap.replication}
\title{Bootstrap replication for a function}
\usage{
bootstrap.replication(x, n, sensitivity, epsilon, fun, inputObject, ...)
}
\arguments{
\item{x}{Vector}

\item{n}{Number of observations}

\item{sensitivity}{Sensitivity of the function}

\item{epsilon}{Numeric differential privacy parameter}

\item{fun}{Function to evaluate}

\item{inputObject}{the Bootstrap mechanism object on which the input function will be evaluated}
}
\value{
Value of the function applied to one bootstrap sample
}
\description{
Bootstrap replication for a function
}
