% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism-bootstrap.R
\name{bootstrapReplication}
\alias{bootstrapReplication}
\title{Bootstrap replication for a function}
\usage{
bootstrapReplication(x, n, sensitivity, epsilon, fun)
}
\arguments{
\item{x}{Vector}

\item{n}{Number of observations}

\item{sensitivity}{Sensitivity of the function}

\item{epsilon}{Numeric differential privacy parameter}

\item{fun}{Function to evaluate}
}
\value{
Value of the function applied to one bootstrap sample
}
\description{
Bootstrap replication for a function
}
