% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{buildTree}
\alias{buildTree}
\title{Recursively build tree object up to .self$depth}
\arguments{
\item{currNode}{Current node, used for recursive call.}
}
\description{
Recursively build tree object up to .self$depth
}
