% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{calculateDepth}
\alias{calculateDepth}
\title{Calculates depth of tree given number of leaves nLeaves.}
\arguments{
\item{nLeaves}{Numeric value indicating the number of leaves in a tree object.}
}
\value{
Numeric value of depth of tree.
}
\description{
Assumes that tree is perfect; i.e. has leaves filled in to complete depth.
Assumes that a tree with only a single root node has depth 1.
}
