% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{calculateIndex}
\alias{calculateIndex}
\title{Calculate index of a node given its parent node.}
\arguments{
\item{isLeftChild}{Boolean. TRUE if finding index of left child, FALSE if finding index of right child.}
}
\value{
Numeric index of child node.
}
\description{
Calculate index of a node given its parent node.
}
