% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{calculateRange}
\alias{calculateRange}
\title{Calculate range}
\arguments{
\item{isLeftChild}{Boolean. TRUE if finding range of left child, FALSE if finding range of right child.}
}
\value{
Numeric of length 2; c(min,max) where min is minumum of child range and max is maximum of child range.
}
\description{
Calculates range of child node given range of parent node. Returned range will be half of parent range;
if left child will be bottom half, if right child will be upper half of parent range.
}
