% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{cdfMedian}
\alias{cdfMedian}
\title{Median of Input Data}
\usage{
cdfMedian(cdf)
}
\arguments{
\item{cdf}{CDF generated with the treePostCDF function.}
}
\value{
Estimate of the median and the proportion that was used to create the estimate.
}
\description{
Estimated median of input data as postprocessed from tree.
}
\details{
Note that the median is just the cdf of a distribution evaluated at the 50th percentile. Given an empirical cdf genderated by the
treePostCDF function, we can report this value according to the noisy counts of the tree exactly if the 50th percentile is included.
If the granularity of the bins and counts of the data at the lowest level of the noisy tree are such that the 50th percentile is
not included in the output cdf, the function will instead output the closest count.

E.g. if the input has

cdf$proportions <- c(0, 0.3. 0.4, 0.7, 1.0),
cdf$counts <- c(0,3,4,7,10)

then the function will output the cdf count associated with 0.4 as an estimate. The output is a tuple of both the estimated median, 
med$val, and the value of the proportion that was used to create the estimate.
}
