% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{check_histogram_bins}
\alias{check_histogram_bins}
\title{Check histogram bins argument}
\usage{
check_histogram_bins(n_bins, n)
}
\arguments{
\item{n_bins}{The number of cells in which to tabulate values.}

\item{n}{A numeric vector of length one specifying the number of
observations in in the data.}
}
\value{
Number of bins
}
\description{
Utility function to check bins argument to histogram. If number of bins 
   is not provided, the Sturges method is used.
}
