% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_histogram_mechanism}
\alias{check_histogram_mechanism}
\title{Utility function to verify the type of histogram mechanism}
\usage{
check_histogram_mechanism(mechanism)
}
\arguments{
\item{mechanism}{Character string specifying the mechanism

Verifies that the mechanism is one of `noisy`, `stability`, or `random` and returns 
the mechanism if so, else throws an error}
}
\description{
Utility function to verify the type of histogram mechanism
}
\examples{
check_histogram_mechanism('stability')
}

