% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_histogram_n}
\alias{check_histogram_n}
\title{Utility function to check sufficient n}
\usage{
check_histogram_n(accuracy, n, n_bins, epsilon, delta, alpha)
}
\arguments{
\item{accuracy}{The accuracy we need to guarantee}

\item{n}{Number of observations}

\item{n_bins}{Number of cells in which to tabulate values}

\item{epsilon}{Numeric, epsilon parameter for differential privacy}

\item{delta}{something here}

\item{alpha}{something here}
}
\value{
something here
}
\description{
Utility function to check sufficient n
}
