% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-validation.R
\name{check_n_validity}
\alias{check_n_validity}
\title{Check validity of n}
\usage{
check_n_validity(n)
}
\arguments{
\item{n}{the input n from te user}
}
\value{
n, if n is a positive whole number
}
\description{
n should always be a positive whole number, check the user's input
}
