% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkAccuracy}
\alias{checkAccuracy}
\title{Utility function for checking that accuracy is acceptably defined.}
\usage{
checkAccuracy(accuracy)
}
\arguments{
\item{accuracy}{}
}
\value{
accuracy or errors.
}
\description{
Verifies accuracy is greater than 0.
}
