% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-covariance.R
\name{checkAccuracyVals}
\alias{checkAccuracyVals}
\title{Check validity of accuracyVals}
\usage{
checkAccuracyVals(accuracyVals, expectedLength)
}
\arguments{
\item{accuracyVals}{Vector of accuracy values where the ith accuracy will be used for the ith covariance calculation in 
flattened lower triangle of covariance matrix.}

\item{expectedLength}{Integer value. Specifies how long the output ought to be.}
}
\value{
accuracyVals or errors.
}
\description{
Verifies that accuracyVals has expected length and contains no values less than or equal to 0.
}
