% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkEmpty}
\alias{checkEmpty}
\title{Checks if input n has NULL or NA value.
Errors if emptyOkay is FALSE.}
\usage{
checkEmpty(n, emptyOkay = FALSE)
}
\arguments{
\item{n}{Arbitrary input n}

\item{emptyOkay}{Boolean. If FALSE, function will error if n is NULL or NA. If TRUE, no error is raised.}
}
\value{
n
}
\description{
Checks if input n has NULL or NA value.
Errors if emptyOkay is FALSE.
}
