% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkEpsilon}
\alias{checkEpsilon}
\title{Epsilon Parameter Check}
\usage{
checkEpsilon(epsilon, multipleEps = FALSE, expectedLength = 1)
}
\arguments{
\item{epsilon}{A vector, that ought to be positive.}

\item{multipleEps}{A boolean value. If TRUE, multiple epsilon paramaters may be defined. Default to FALSE.}

\item{expectedLength}{Integer value. Specifies how long the output ought to be.}
}
\value{
The supplied epsilon if acceptable, otherwise an error 
   message interupts.
}
\description{
Utility function for checking that epsilon is acceptably defined.
}
