% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{checkHistogramVariableType}
\alias{checkHistogramVariableType}
\title{Check the histogram variable type entered by the user}
\usage{
checkHistogramVariableType(varType)
}
\arguments{
\item{varType}{The variable type of the data that was entered by the user}
}
\value{
No return value, will only send an error message if variable type is invalid.
}
\description{
The variable type for a histogram must be 'numeric', 'integer', 'logical', or 'character'.
If it is not one of these, send an error message to the user.
}
