% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-validation.R,
%   R/utilities-input-validation.R
\name{checkImputationRange}
\alias{checkImputationRange}
\title{Error check imputation range for numeric or integer variables}
\usage{
checkImputationRange(imputationRange, rng, varType)

checkImputationRange(imputationRange, rng, varType)
}
\arguments{
\item{imputationRange}{The imputation range entered by the user}

\item{rng}{The data range entered by the user}

\item{varType}{The variable type for the histogram data}

\item{imputationRange}{The imputation range entered by the user}

\item{rng}{The data range entered by the user}

\item{varType}{The variable type for the histogram data}

\item{imputationRange}{The imputation range entered by the user}

\item{rng}{The data range entered by the user}

\item{varType}{The variable type for the histogram data}
}
\value{
the imputation range that will be used for `fillMissing()`.

the imputation range that will be used for `fillMissing()`.
Error check imputation range for numeric or integer variables

Check that the entered imputation range is within the entered data range. If yes, return
the entered imputation range, which will be used as the imputation range for the call
to the utility function `fillMissing()`. If not, return the data range. 
If the imputation range is NULL, default to the data range.

We check if the imputation range is within the data range because it is a privacy concern.
If the imputation range is outside of the data range, NA values will be replaced with values 
outside of the data range, which will show that there are NA values in the data or skew the 
result when the differentially private estimate is released.

the imputation range that will be used for `fillMissing()`.
}
\description{
Check that the entered imputation range is within the entered data range. If yes, return
the entered imputation range, which will be used as the imputation range for the call
to the utility function `fillMissing()`. If not, return the data range. 
If the imputation range is NULL, default to the data range.

Check that the entered imputation range is within the entered data range. If yes, return
the entered imputation range, which will be used as the imputation range for the call
to the utility function `fillMissing()`. If not, return the data range. 
If the imputation range is NULL, default to the data range.
}
\details{
We check if the imputation range is within the data range because it is a privacy concern.
If the imputation range is outside of the data range, NA values will be replaced with values 
outside of the data range, which will show that there are NA values in the data or skew the 
result when the differentially private estimate is released.

We check if the imputation range is within the data range because it is a privacy concern.
If the imputation range is outside of the data range, NA values will be replaced with values 
outside of the data range, which will show that there are NA values in the data or skew the 
result when the differentially private estimate is released.
}
