% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkLength}
\alias{checkLength}
\title{Check if input xs has length n}
\usage{
checkLength(xs, n)
}
\arguments{
\item{xs}{A vector, factor, or object for which length is defined.}

\item{n}{A numeric value of length 1. The expected length of xs.}
}
\value{
xs or error.
}
\description{
Check if input xs has length n
}
