% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkN}
\alias{checkN}
\title{Check validity of n}
\usage{
checkN(n)
}
\arguments{
\item{n}{the input n from the user}
}
\value{
n, if n is a positive whole number
}
\description{
n should always be a positive whole number, check the user's input
}
