% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkN1D}
\alias{checkN1D}
\title{Helper function for helpN}
\usage{
checkN1D(n, emptyOkay = FALSE)
}
\arguments{
\item{n}{the input n (often number of data points in data set) from the user}

\item{emptyOkay}{A boolean. True if NA or NULL values are allowed, FALSE otherwise.}
}
\value{
n, if n is a series of positive integers with expected length and only containing NAs or NULL values if allowed.
}
\description{
Checks that n is always a positive whole number; n may be NA or NULL only if emptyOkay=TRUE.
}
