% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{checkNLeaves}
\alias{checkNLeaves}
\title{Checks if number of leaves is a power of two or not}
\arguments{
\item{nLeaves}{Numeric value indicating the number of leaves in a tree object.}
}
\value{
Boolean value. TRUE if nLeaves is a power of two. Stops and returns error otherwise.
}
\description{
Checks if number of leaves is a power of two or not
}
