% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkNumeric1D}
\alias{checkNumeric1D}
\title{Check if input n is numeric.
Generally errors if n not numeric.
Errors for NULL or NA values if emptyOkay=FALSE.}
\usage{
checkNumeric1D(n, emptyOkay = FALSE)
}
\arguments{
\item{n}{Arbitrary length 1 input}

\item{emptyOkay}{If FALSE, function will error if n is NULL or NA. If TRUE, no error is raised.}
}
\value{
n
}
\description{
Check if input n is numeric.
Generally errors if n not numeric.
Errors for NULL or NA values if emptyOkay=FALSE.
}
