% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkRange}
\alias{checkRange}
\title{Range Parameter Check}
\usage{
checkRange(rng, varType)
}
\arguments{
\item{rng}{Range. A numeric vector of one of two formats:
1. Of length two, that ought to be an ordered pair, representing 
the maximum and minimum bounds on the data
2. A sequence of ordered pairs in columns, where each row represents
 the maximum and minimum bounds on some subsets of the data (e.g. of different data columns)}

\item{varType}{The variable type; e.g. 'logical', 'integer', 'numeric', 'character'.}
}
\value{
An ordered pair.
}
\description{
Checks if a supplied range is an ordered pair.
}
\examples{

checkRange(c(1,3))
checkRange(1:3)s
\dontrun{checkRange(1)}
}
