% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkRange1D}
\alias{checkRange1D}
\title{One-Dimensional Range Parameter Check}
\usage{
checkRange1D(rng, varType)
}
\arguments{
\item{rng}{Range. A numeric vector that ought to be an ordered pair.}

\item{varType}{String describing the variable type. One of 'logical' or 'numeric'.}
}
\description{
Helper function for checkRange. Checks if a supplied range is an ordered pair.
Coerces any vector of length two or greater into an ordered pair, and issues 
an error for shorter vectors.
}
