% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-validation.R
\name{checkVariableType}
\alias{checkVariableType}
\title{Checking variable types}
\usage{
checkVariableType(type, in_types)
}
\arguments{
\item{type}{A character specifying the type of the variable.}

\item{in_types}{A vector of acceptable types of variables.}
}
\value{
The original character string indicating the variable type.
}
\description{
Verifies that the variable is an element in the set of acceptable types.
}
\examples{

checkVariableType(type='Numeric', in_types=c('Numeric', 'Factor'))
}
