% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{coefficient.release}
\alias{coefficient.release}
\title{Release additional model coefficients from DP covariance matrix}
\usage{
coefficient.release(formula, release, n)
}
\arguments{
\item{formula}{Formula, regression formula used on data}

\item{release}{Numeric, private release of covariance matrix}

\item{n}{Integer, indicating number of observations}
}
\description{
Function to extract regression coefficients using the differentially private covariance matrix 
   via the sweep operator. This is the function to obtain coefficients for additional models
   after the \code{covariance.release()} function has already been called.
}

