% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-covariance.R
\name{coefficientRelease}
\alias{coefficientRelease}
\title{Release additional model coefficients from DP covariance matrix}
\usage{
coefficientRelease(formula, release, n)
}
\arguments{
\item{formula}{Formula, regression formula used on data}

\item{release}{Numeric, private release of covariance matrix}

\item{n}{Integer, indicating number of observations}
}
\description{
Function to extract regression coefficients using the differentially private covariance matrix 
via the sweep operator. This is the function to obtain coefficients for additional models
after the \code{covariance.release()} function has already been called.
}
\examples{
\dontrun{
range.income <- c(-10000, 713000)
range.education <- c(1, 16)
range <- rbind(range.income, range.education)
dpCov <- dpCovariance$new(mechanism="mechanismLaplace",var.type = 'numeric', n = 10000,
                          epsilon = 1, columns = c("income", "educ"), rng = range, formula='income~educ')
out <- dpCov$release(PUMS5extract10000)
coefficient.release('educ~income', out$release, n=10000)
}
}
