% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{countAbove}
\alias{countAbove}
\title{Recursively compute counts from above}
\usage{
countAbove(tree, countsBelow, wAboves)
}
\arguments{
\item{tree}{Tree, formatted as a list of arrays where the contents of the ith array in the list is the ith level of the tree.}

\item{countsBelow}{Array of counts of same length as tree, as calculated by countBelow function}

\item{wAboves}{Weights computed from above, assuming each node in tree has same variance.}
}
\value{
List of counts for each node of the tree.
}
\description{
See extra_docs/tree-post-processing for formula. This assumes that the variance is the same for every node at in the tree.
}
\examples{
t <- list(c(10), c(6,4), c(3,3,1,3))
wB <- wBelow(t)
wA <- wAbove(t, wB)
cB <- countBelow(t, wB)
cA <- countAbove(t, cB, wA) #t and cA should be equal, since counts in t have no added noise.
}
