% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{countBelow}
\alias{countBelow}
\title{Recursively compute counts from below}
\usage{
countBelow(tree, wBelows)
}
\arguments{
\item{tree}{Tree, formatted as a list of arrays where the contents of the ith array in the list is the ith level of the tree.}

\item{wBelows}{Array of weights of same length as tree, where the ith weight corresponds to the ith weight calculated from below.}
}
\value{
List of counts for each node of the tree.
}
\description{
See extra_docs/tree-post-processing for formula. This assumes that the variance is the same for every node at in the tree.
}
\examples{
t <- list(c(10), c(6,4), c(3,3,1,3))
w <- wBelow(t)
c <- countBelow(t, w) # t and c should be equal, since counts in t have no added noise.
}
