% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{covariance.formatRelease}
\alias{covariance.formatRelease}
\title{Function to convert unique private covariances into symmetric matrix}
\usage{
covariance.formatRelease(release, intercept, columns)
}
\arguments{
\item{release}{Numeric private release of elements in lower triangle of covariance matrix}

\item{intercept}{Logical indicating whether the intercept is included}

\item{columns}{Character vector with column names

This function is used by the mechanism in post-processing and not intended for interactive post-processing}
}
\description{
Function to convert unique private covariances into symmetric matrix
}
