% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{covariance.postLinearRegression}
\alias{covariance.postLinearRegression}
\title{Function to perform linear regression using private release of covariance matrix}
\usage{
covariance.postLinearRegression(release, n, intercept, formulae)
}
\arguments{
\item{release}{Numeric private release of covariance matrix}

\item{n}{Integer indicating number of observations}

\item{intercept}{Logical indicating whether the intercept is included}

\item{formulae}{List of R formulae}
}
\value{
something here
}
\description{
Function to perform linear regression using private release of covariance matrix
}
