% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{covariance.release}
\alias{covariance.release}
\title{Function to release a differentially private inner product of input matrix.}
\usage{
covariance.release(x, var.type, n, epsilon, rng, columns, delta = 1e-06,
  intercept = FALSE, formulae = NULL, mechanism = "laplace")
}
\arguments{
\item{x}{A numeric data frame with at least two columns.}

\item{var.type}{A character vector specifying variable type of \code{x}. 
Should be of length one and should contain either 'numeric', 
'logical', or 'integer'.}

\item{n}{A numeric vector of length one specifying the number of
observations in \code{x}.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{rng}{A numeric matrix of 2-tuples with the lower and upper bounds for
each of P variables in \code{x}, dimensions Px2.}

\item{columns}{A character vector indicating columns in \code{x} to be 
included in the covariance matrix. Length should be equal to the number 
of columns the user wants to include.}

\item{delta}{A numeric vector representing the probability of an arbitrary
leakage of information from \code{x}. Should be of length one 
and should be a very small value. Default to 10^-6.}

\item{intercept}{A logical vector of length one indicating whether an 
intercept should be added prior to evaluating the inner product x'x.
Default to FALSE.}

\item{formulae}{The regression equations the user would like to perform on
the covariance matrix. The equations should be of class 'formula'. The 
user may specify as many equations as desired. Default to FALSE.}

\item{mechanism}{A character vector specifying the mechanism used to apply 
noise to the covariance matrix. Should be of length one and contain 
either 'laplace', 'gaussian', or 'wishart'. Default to 'laplace'.}
}
\value{
Differentially private covariance matrix of \code{x}.
}
\description{
Function to release a differentially private inner product of input matrix.
}
\examples{

data(PUMS5extract10000, package = "PSIlence")
data <- data.frame(income = PUMS5extract10000$income, education = PUMS5extract10000$educ)
range.income <- c(-10000, 713000)
range.education <- c(1, 16)
range <- rbind(range.income, range.education)
covariance.release(x = data, var.type = 'numeric', n = 10000, epsilon = 0.2, rng = range, 
                   columns = c("income", "education"), formulae = income ~ education)
}
