% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{covariance.sensitivity}
\alias{covariance.sensitivity}
\title{Function to get the sensitivity of the covariance matrix}
\usage{
covariance.sensitivity(n, rng, intercept)
}
\arguments{
\item{n}{Integer indicating the number of observations}

\item{rng}{Numeric matrix of 2-tuples with the lower and upper bounds for each of P variables, dimensions Px2}

\item{intercept}{Logical indicating whether an intercept should be included}
}
\description{
Function to get the sensitivity of the covariance matrix
}
