% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-covariance.R
\name{covarianceFormatRelease}
\alias{covarianceFormatRelease}
\title{Function to convert unique private covariances into symmetric matrix}
\usage{
covarianceFormatRelease(release, columns)
}
\arguments{
\item{release}{Differentially private release of elements in lower triangle 
of covariance matrix.}

\item{columns}{A character vector indicating columns in the private 
covariance to be included in the output. Length should be equal to the 
number of columns the user wants to include.}
}
\value{
A symmetric differentially private covariance matrix.

This function is used by the mechanism in post-processing and not intended for interactive post-processing
}
\description{
Function to convert unique private covariances into symmetric matrix
}
