% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-covariance.R
\name{covarianceGetAccuracy}
\alias{covarianceGetAccuracy}
\title{Function to determine accuracy of dpCovariance for a given epsilon}
\usage{
covarianceGetAccuracy(epsilonPart, sens, alpha = 0.05)
}
\arguments{
\item{epsilonPart}{A numeric vector of length one. The amount of the global epsilon that is 
allocated to each covariance calculation within the covariance matrix.}

\item{sens}{A numeric vector of length one. The sensitivity of the covariance function}

\item{alpha}{A numeric vector of length one specifying the numeric 
statistical significance level. Default to 0.05.}
}
\value{
Accuracy
}
\description{
Since the covariance matrix that is released by the dpCovariance statistic has
the global epsilon budget distributed between each of the calculated covariances,
this function is passed the amount of epsilon that each individual covariance
calculation receives, rather than the global epsilon.
}
