% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sampleandaggregate.R
\name{create_partition_of_indices}
\alias{create_partition_of_indices}
\title{Generates a list of lists, where each individual list is a set of indices,
each individual list is disjoint, and the union over the set of all lists
is the set of all indices. That is, it returns a partition of indices.}
\usage{
create_partition_of_indices(n_obs, n_subsets)
}
\arguments{
\item{n_obs}{the number of observations in the data set}

\item{n_subsets}{the number of desired subsets}
}
\value{
a list of lists, where each list is a set of indices representing one of the subsets of the partition
}
\description{
Generates a list of lists, where each individual list is a set of indices,
each individual list is disjoint, and the union over the set of all lists
is the set of all indices. That is, it returns a partition of indices.
}
