% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-categorical.R
\name{createfields}
\alias{createfields}
\title{Create plot fields}
\usage{
createfields(v, r, varName)
}
\arguments{
\item{v}{an empty list needing initialization}

\item{r}{a release object for that variable}

\item{varName}{name of variable}
}
\description{
Designate the plot type based on the variable type.
Also assign the plot variable based on the variable the plot is for.
}
\details{
This function is called by \code{\link{release2json}} in \code{\link{utilities-postprocessing.R}}.
}
