% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-postprocessing.R
\name{createJSON}
\alias{createJSON}
\title{Function to create JSON file defining differentially private statistics}
\usage{
createJSON()
}
\description{
Function to create JSON file defining differentially private statistics
}
