% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{determineBins}
\alias{determineBins}
\title{Determine Bins}
\usage{
determineBins(var.type, rng, bins, n.bins, impute, granularity, object)
}
\arguments{
\item{var.type}{Character, the variable type.}

\item{rng}{Numeric, a priori estimate of the lower and upper bounds of a
variable taking numeric values. Ignored for categorical types. Maybe be null for 
numeric or integer types, in which case the stability mechanism is used.}

\item{bins}{Character or numeric, the available bins or levels of a variable. Character 
for categorical variables, a vector of numbers for numeric variables.}

\item{n.bins}{Integer, the number of bins to release.}

\item{impute}{Boolean, if true then the mechanism should replace missing values with known 
values from the data.If false, the mechanism should leave missing values as `NA`}

\item{granularity}{Numeric, the width of each histogram bin.}

\item{object}{Object, the dpHistogram object for the given variable (used it access and assign variable type)}
}
\value{
a vector of histogram bins. Character vector for categorical variables. Numeric 
   vector for logical, numeric, and integer variables.
}
\description{
Determine the bins of the histogram based on the inputs from the user
}
