% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{determineMechanism}
\alias{determineMechanism}
\title{Determine Mechanism}
\usage{
determineMechanism(varType, rng, bins, nBins, granularity)
}
\arguments{
\item{varType}{Character, the variable type.}

\item{rng}{Numeric, a priori estimate of the lower and upper bounds of a
variable taking numeric values. Ignored for categorical types. Maybe be 
null for numeric or integer types, in which case the stability mechanism is used.}

\item{bins}{Character or numeric, the available bins or levels of a variable. 
Character for categorical variables, a vector of numbers for numeric variables.}

\item{nBins}{Integer, the number of bins to release.}

\item{granularity}{Numeric, the width of each histogram bin.}
}
\value{
a string indicating the mechanism to use when constructing the differentially private histogram
}
\description{
This is a set of helper functions to determine which mechanism to use when
calculating the histogram (either the stability mechanism or the Laplace
mechanism).
}
