% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-composition.R
\name{distributeEpsilon}
\alias{distributeEpsilon}
\title{Epsilon distribution between subcalculations.}
\usage{
distributeEpsilon(globalEps, nCalcs = NULL, epsilonDist = NULL)
}
\arguments{
\item{globalEps}{Global epsilon budget for all calculations}

\item{nCalcs}{Integer. The number of calculations to split the global epsilon budget between.}

\item{epsilonDist}{Optional numeric vector of percentages (between 0 and 1) which, if specified, should be of length nCalcs,
describing how to to destribute the globalEps between each of the nCalcs.}
}
\value{
Portion of epsilon to be budgeted to each individual covariance calculation in the covariance matrix.
}
\description{
Determines the amount of epsilon to budget for some collection of calculations of length \code{nCalcs}
given either only the global epsilon budget \code{globalEps} and the number of sub-calculations to be computed,
\code{nCalcs}, or these plus an additional parameter, \code{epsilonDist}, which describes how the global epsilon
ought to be distributed between the \code{nCalcs}.
}
\details{
If \code{epsilonDist} is not specified, \code{globalEps} is budgeted evenly between each sub-calculation.

If \code{epsilonDist} is specified, \code{globalEps} is budgeted according to this.

Budgeting is performed by simple composition of epsilon differential privacy, which states that if a series of calculations
\eqn{f_1, \ldots, f_n} have privacy budgets of \eqn{\epsilon_1 \ldots \epsilon_n} respectively, then the computation 
\eqn{f = (f_1, \ldots, f_n)} will have privacy budget \eqn{\epsilon = \sum_{i=1}^n \epsilon_i}.
}
\references{
C. Dwork, A. Roth The Algorithmic Foundations of Differential Privacy, Chapter 3.5 Theorem 3.14
Composition Theorems p.42. August 2014.
}
