% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{dp.covariance}
\alias{dp.covariance}
\title{Differentially Private Covariance Matrix}
\usage{
dp.covariance(x, n, rng, epsilon, columns, intercept, formulae)
}
\arguments{
\item{x}{A numeric data frame with at least two columns.}

\item{n}{A numeric vector of length one specifying the number of
observations in \code{x}.}

\item{rng}{A numeric matrix of 2-tuples with the lower and upper bounds for
each of P variables in \code{x}, dimensions Px2.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{columns}{A character vector indicating columns in \code{x} to be 
included in the covariance matrix. Length should be equal to the number 
of columns the user wants to include.}

\item{intercept}{A logical vector of length one indicating whether an 
intercept should be added prior to evaluating the inner product x'x.}

\item{formulae}{The regression equations the user would like to perform on
the covariance matrix. The equations should be of class 'formula'. The 
user may specify as many equations as desired.}
}
\value{
A list with fields `name` specifying the statistic and `stat` with 
   the lower triangle of the covariance matrix.
}
\description{
Function to evaluate the covariance matrix and specify parameters for covariance functions.
}
