% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{dp.heavyhitters}
\alias{dp.heavyhitters}
\title{Function to evaluate most common values and specify arguments to post-processing}
\usage{
dp.heavyhitters(x, var.type, epsilon, delta, n, k, sensitivity)
}
\arguments{
\item{x}{Vector of categorical data (character, factor)}

\item{var.type}{Character string indicating data type}

\item{epsilon}{Epsilon value for differential privacy}

\item{n}{Integer indicating number of observations}

\item{k}{Integer querying the most common \code{k} categories}

\item{sensitivity}{Numeric the sensitivity of the statistic}
}
\value{
List with true value of statistic and parameters to be passed to post-processing
}
\description{
Function to evaluate most common values and specify arguments to post-processing
}
