% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{dp.mean}
\alias{dp.mean}
\title{DP Mean}
\usage{
dp.mean(x, var.type, n, sensitivity, epsilon)
}
\arguments{
\item{x}{A numeric, logical, or integer vector.}

\item{var.type}{A character vector specifying variable type of \code{x}. 
Should be of length one and should contain either 'numeric', 
'logical', or 'integer'.}

\item{n}{A numeric vector of length one specifying the number of
observations in \code{x}.}

\item{sensitivity}{The difference of the range of the data divided 
by \code{n}.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}
}
\value{
A list with fields `name` specifying the statistic and `stat` with 
   the value of the statistic.
}
\description{
Function to evaluate the mean and specify parameters for mean functions.
}
