% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-glm.R
\name{dp.poisson}
\alias{dp.poisson}
\title{Differentially private objective function for Poisson regression}
\usage{
dp.poisson(n, epsilon, formula, intercept)
}
\arguments{
\item{n}{Integer indicating number of observations}

\item{epsilon}{Numeric epsilon parameter for differential privacy}

\item{formula}{Formula for the Poisson regression model}

\item{intercept}{Logical indicating whether the intercept should be added to the model}
}
\description{
Differentially private objective function for Poisson regression
}
