% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-tree.R
\name{dp.tree}
\alias{dp.tree}
\title{Evaluate a differentially private binary tree}
\usage{
dp.tree(x, var.type, n, rng, epsilon, sensitivity, gran, variance, percentiles)
}
\arguments{
\item{x}{Vector of numeric observations}

\item{var.type}{Character vector specifying the variable type, should be
one of \code{c('numeric', 'integer')}}

\item{n}{Number of observations}

\item{rng}{An a priori estimate of the range of \code{x}}

\item{epsilon}{Privacy parameter epsilon, should be between zero and one}

\item{sensitivity}{The sensitivity of the statistic}

\item{gran}{The granularity at which \code{x} is represented in the tree}

\item{variance}{The variance of the noise used to perturb tree nodes}

\item{percentiles}{Vector of percentiles used in the post-processing
quantile function}
}
\value{
List of values including the true value of the binary tree and the associated data
     needed for the noisy release and post-processing
}
\description{
Evaluate a differentially private binary tree
}
