% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-heavyhitters.R
\docType{class}
\name{dpHeavyHitters-class}
\alias{dpHeavyHitters-class}
\alias{dpHeavyHitters}
\title{Differentially private heavy hitters}
\arguments{
\item{mechanism}{Character, the mechanism used to perturb the output.}

\item{var.type}{Character, the variable type, one of \code{'character'} or
\code{'factor'}.}

\item{variable}{Character, the variable name in the data frame.}

\item{n}{Integer, the number of observations.}

\item{epsilon}{Numeric, the privacy loss parameter.}

\item{k}{Integer, the number of bins to release.}

\item{bins}{Character, the available bins, or the levels of the categorical variable.}

\item{alpha}{Numeric, level of statistical significance, default 0.05.}

\item{delta}{Numeric, probability of privacy loss beyond \code{epsilon}.}
}
\description{
Differentially private heavy hitters
}

