% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-glm.R
\name{dpLogit}
\alias{dpLogit}
\title{Differentially private objective function for Logistic regression}
\usage{
dpLogit()
}
\value{
List with the name and objective function
Differentially private objective function for Probit regression

List with the name and objective function
Differentially private objective function for Logistic regression

List with the name and objective function
}
\description{
Differentially private objective function for Logistic regression
}
