% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mean.R
\name{dpMeanPostProcess}
\alias{dpMeanPostProcess}
\title{Post-processing on differentially private mean, called within the \code{.self$mechanism$evaluate}
function, which in turn is called within \code{dpMean$release}.}
\arguments{
\item{out}{Input dataframe that describes the differentially private release that was created in
\code{.self$mechanism$evaluate}. This dataframe will have at least one pre-existing attribute, 
out$release, which is a numeric value of length one that is the differentially private mean.

This function is able to calculate a confidence interval on the output. If the variable was logical,
standard deviation, median, and a histogram may be additionally computed as post-processing steps.

Additionally, known portions of the input such as the variable name 
and the epsilon value may be appended with no extra privacy loss.}
}
\value{
Dataframe \code{out}, updated to include post-processed values.
}
\description{
Post-processing on differentially private mean, called within the \code{.self$mechanism$evaluate}
function, which in turn is called within \code{dpMean$release}.
}
