% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mean.R
\name{dpMeanRelease}
\alias{dpMeanRelease}
\title{Differentially private mean release}
\arguments{
\item{data}{Dataframe with a column named .self$variable, where
 that column has data of type .self$varType and which is bounded by 
 .self$rng.

Assigns to .self$result a dataframe that describes the differentially private 
mean, calculated by some mechanism as defined in .self$mechanism, of that
column of the dataframe and any post-processing on that output. This 
postprocessing is done in @seealso{dpMean$postProcess}

Note that the actual differentially private release is calculated in a call to the
differentially private mechanism .self$mechanism's \code{evaluate} function within 
the \code{dpMean$release} function.}
}
\description{
Differentially private mean release
}
