% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dpNoise}
\alias{dpNoise}
\title{Differentially Private Noise Generator}
\usage{
dpNoise(n, scale, dist, shape = NULL, seed = NULL)
}
\arguments{
\item{n}{An integer giving number of variates needed.}

\item{scale}{Numeric, the scale for the distribution.}

\item{dist}{A character specifying the distribution from which to draw the 
noise.}

\item{shape}{An integer giving the shape parameter for the gamma
distribution. Default to \code{NULL}.}

\item{seed}{An integer indicating a seed for R's PNRG, defaults 
to \code{NULL}.}
}
\value{
Cryptographically secure noise vector or matrix.
}
\description{
Compile noise from a cryptographically secure random variates to achieve
   differentially private statistics.
}
\examples{

laplace_noise <- dpNoise(n=1000, scale=1, dist='laplace')
gaussian_noise <- dpNoise(n=1000, scale=1, dist='gaussian')
laplace_noise_repeatable <- dpNoise(n=1, scale=1, dist='laplace', seed=96845)
}
\seealso{
\code{\link{dpUnif}}
}

