% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-glm.R
\name{dpOLS}
\alias{dpOLS}
\title{Differentially private objective function for linear regression}
\usage{
dpOLS()
}
\value{
List with the name and objective function
}
\description{
Differentially private objective function for linear regression
}

