% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-glm.R
\name{dpPoisson}
\alias{dpPoisson}
\title{Differentially private objective function for Poisson regression}
\usage{
dpPoisson()
}
\value{
List with the name and objective function
}
\description{
Differentially private objective function for Poisson regression
}
