% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-glm.R
\name{dpProbit}
\alias{dpProbit}
\title{Differentially private objective function for Probit regression}
\usage{
dpProbit()
}
\value{
List with the name and objective function
}
\description{
Differentially private objective function for Probit regression
}

