% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-sampleandaggregate.R
\docType{class}
\name{dpSampleAndAggregate-class}
\alias{dpSampleAndAggregate-class}
\alias{dpSampleAndAggregate}
\title{Sample and Aggregate "Statistic"}
\arguments{
\item{inner_fun}{Character, the function calculated within each subset of the partition.}

\item{aggregation_fun}{Character, the function calculated over the set of results from
each subset of the partition.}

\item{mechanism}{Character, the privacy mechanism. For \code{dpSampleAndAggregate}, one
of \code{c('mechanismLaplace')}.}

\item{n_subsets, }{Numeric, the number of subsets desired in the partition}

\item{varType}{Character, the R variable type. One of \code{c('numeric',
'integer', 'logical')}.}

\item{variable}{Character, variable name.}

\item{n}{Integer, number of observations}

\item{inner_fun_sens}{Numeric, sensitivity of inner function.}

\item{aggregation_fun_sens}{Numeric, sensitivity of aggregation function.}

\item{subset_size}{Numeric, minimum size of a subset of the partition.}

\item{rng}{Numeric, a priori estimate of the range}

\item{epsilon}{Numeric, privacy cost parameter}

\item{accuracy}{Numeric, accuracy guarantee given \code{epsilon}}

\item{imputeRng}{Numeric, range within which missing values are imputed. If \code{NULL},
the range provided in \code{rng} is used.}

\item{alpha}{Numeric, the level of statistical significance. Default 0.05.}
}
\description{
Sample and Aggregate "Statistic"
}

