% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\docType{class}
\name{dpTree-class}
\alias{dpTree-class}
\alias{dpTree}
\title{Differentially private binary tree}
\arguments{
\item{mechanism}{Character, the privacy mechanism.}

\item{varType}{Character, the R variable type. One of \code{'numeric'} or
\code{'integer'}.}

\item{Variable}{Character, variable name.}

\item{n}{Integer, number of observations.}

\item{rng}{Numeric, a priori estimate of the range.}

\item{gran}{Numeric, the granularity of the variable.}

\item{epsilon}{Numeric, privacy cost parameter.}

\item{accuracy}{Numeric, accuracy guarantee given \code{epsilon}.}

\item{imputeRng}{Numeric, range within which missing values are imputed. If \code{NULL},
the range provided in \code{rng} is used.}

\item{percentiles}{Numeric, the percentiles to evaluate in post-processing. Optional, 
default \code{NULL}.}

\item{alpha}{Numeric, the level of statistical significance. Default 0.05.}
}
\description{
Differentially private binary tree
}

