% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dpUnif}
\alias{dpUnif}
\title{Draw cryptographically secure random variates from uniform distribution}
\usage{
dpUnif(n, seed = NULL)
}
\arguments{
\item{n}{Integer giving number of variates needed}

\item{seed}{Integer indicating a seed for R's PNRG, defaults to \code{NULL}

Draws secure random variates from the uniform distribution through \code{openssl}.
If a seed is provided, the \code{runif} function is used to draw the random variates.}
}
\description{
Draw cryptographically secure random variates from uniform distribution
}
\examples{
uniform_secure <- dpUnif(n=1000)
uniform_repeatable <- dpUnif(n=1, seed=75436)
}
