% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-variance.R
\name{dpVariancePostProcess}
\alias{dpVariancePostProcess}
\title{Post-processing on differentially private variance, called within the \code{.self$mechanism$evaluate}
function, which in turn is called within \code{dpVariance$release}.}
\arguments{
\item{out}{Input dataframe that describes the differentially private release that was created in
\code{.self$mechanism$evaluate}. This dataframe will have at least one pre-existing attribute,
out$release, which is a numeric value of length one that is the differentially private variance.

This function then post-processes the standard deviation from the variance and adds it to the dataframe
as \code{out$std}, as long as the variance is positive.

Additionally, known portions of the input such as the variable name and the accuracy
of the DP variance and the epsilon value may be appended with no extra privacy loss.}
}
\value{
Dataframe \code{out}, updated to include post-processed values.
}
\description{
Post-processing on differentially private variance, called within the \code{.self$mechanism$evaluate}
function, which in turn is called within \code{dpVariance$release}.
}
