% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{estAbove}
\alias{estAbove}
\title{Function to estimate the nodes of a tree using noisy parent and adjacent nodes}
\usage{
estAbove(w, tree, idx, parent, adjacent)
}
\arguments{
\item{w}{Weight used construct the estimate}

\item{tree}{Data frame with binary tree attributes and node values}

\item{idx}{Index of the node for which the estimate is evaluated}

\item{parent}{Index of the parnet node}

\item{adjacent}{Index of the adjacent node}
}
\value{
Noisy node estimate
}
\description{
Function to estimate the nodes of a tree using noisy parent and adjacent nodes
}
