% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{estBottomUp}
\alias{estBottomUp}
\title{Function to estimate a noisy binary tree from the terminal nodes}
\usage{
estBottomUp(tree, terminalLevelIndex, nNodes, sigma, invSigmaSq)
}
\arguments{
\item{tree}{Data frame with binary tree attributes and node values}

\item{terminalLevelIndex}{Index of the first terminal leaf node}

\item{nNodes}{Number of nodes in the binary tree}

\item{sigma}{Standard deviation of the noise used to perturb the estimates}

\item{invSigmaSq}{Inverse variance of the noise used in perturbing nodes}
}
\value{
Bottom-up estimate of noisy binary tree in vector form
}
\description{
Function to estimate a noisy binary tree from the terminal nodes
}
