% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{estEfficient}
\alias{estEfficient}
\title{Function to efficiently estimate the nodes of a tree using all available information in the tree}
\usage{
estEfficient(w, tree, idx, parent, adjacent)
}
\arguments{
\item{w}{Weight used construct the estimate}

\item{tree}{Data frame with binary tree attributes and node values}

\item{idx}{Index of the node for which the estimate is evaluated}

\item{parent}{Index of the parnet node}

\item{adjacent}{Index of the adjacent node}
}
\value{
Efficient noisy node estimate
}
\description{
Function to efficiently estimate the nodes of a tree using all available information in the tree
}
