% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{estTopDown}
\alias{estTopDown}
\title{Function to estimate a noisy binary tree from the top down}
\usage{
estTopDown(tree, n, n.nodes, sigma, inv.sigma.sq)
}
\arguments{
\item{tree}{Data frame with binary tree attributes and node values}

\item{n}{Number of observations in the vector represented by the binary tree}

\item{n.nodes}{Number of nodes in the binary tree}

\item{sigma}{Standard deviation of the noise used to perturb the estimates}

\item{inv.sigma.sq}{Inverse variance of the noise used in perturbing nodes}
}
\value{
Top-down estimate of noisy binary tree in vector form
}
\description{
Function to estimate a noisy binary tree from the top down
}

