% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{extract.indices}
\alias{extract.indices}
\title{Function to obtain indices in data frame for dependent & independent variables from a formula}
\usage{
extract.indices(formula, data, intercept)
}
\arguments{
\item{formula}{Formula}

\item{data}{Data frame, in this case being the data frame of a private covariance matrix}

\item{intercept}{Logical indicating whether the intercept is included}
}
\value{
something here
}
\description{
Function to obtain indices in data frame for dependent & independent variables from a formula
}
