% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fillfields}
\alias{fillfields}
\title{Fill in any fields available from release}
\usage{
fillfields(v, r)
}
\arguments{
\item{v}{a copy of the current metadata for a variable}

\item{r}{the additional released information for the variable}
}
\description{
Fill in any fields available from release
}
