% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fillMissing1D}
\alias{fillMissing1D}
\title{Helper function for fillMissing; fills missing values in one-dimensional array}
\usage{
fillMissing1D(x, var.type, lower = NULL, upper = NULL,
  categories = NULL)
}
\arguments{
\item{x}{Input array of missing values.}

\item{var.type}{Variable type of input array to fillMissing function; affects how scaling occurs. 
Type: one of following strings: 'character', 'factor', 'logical', 'integer', 'numeric'.}

\item{lower}{Lower bound for scaling. Type: numeric. Default NULL.}

\item{upper}{Upper bound for scaling. Type: numeric. Default NULL.}

\item{categories}{List of categories. Type: factor. Default NULL.}
}
\value{
Vector \code{x} with missing values imputed
}
\description{
Imputes uniformly in the range of the provided variable.
}
