% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fillMissing2D}
\alias{fillMissing2D}
\title{Helper function for fillMissing. Fills missing values column-wise for matrix.}
\usage{
fillMissing2D(x, var.type, impute.rng = NULL)
}
\arguments{
\item{x}{Numeric matrix with missing values}

\item{var.type}{Variable type of input array.
Type: one of following strings: 'character', 'factor', 'logical', 'integer', 'numeric'.}

\item{impute.rng}{Px2 matrix where the pth row contains the range
within which the pth variable in x is imputed.}
}
\value{
Matrix \code{x} with missing values imputed
}
\description{
Impute uniformly in the range of the provided variable
}
\seealso{
\code{\link{fillMissing}}
}
