% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-covariance.R
\name{fun.covar}
\alias{fun.covar}
\title{Lower triangle of ovariance matrix}
\usage{
fun.covar(x, columns, intercept)
}
\arguments{
\item{x}{Data frame}

\item{columns}{Columns to include in output}

\item{intercept}{Logical, should the intercept be included?}
}
\description{
This function is called by \code{dpCovariance} and
produces the true value to be perturbed.
}
