% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{fun.heavy}
\alias{fun.heavy}
\title{Heavyhitters function}
\usage{
fun.heavy(x)
}
\arguments{
\item{x}{Vector of categorical values}
}
\value{
Sorted histogram with counts for each level
}
\description{
Function to evaluate most common categorical values
}

