% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-covariance.R
\name{funCovar}
\alias{funCovar}
\title{Lower triangle of covariance matrix}
\usage{
funCovar(x, intercept)
}
\arguments{
\item{x}{Data frame}

\item{intercept}{Logical, should the intercept be included?}

\item{columns}{Columns to include in output}
}
\description{
This function is called by \code{dpCovariance} and
produces the true value to be perturbed.
}
