% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-heavyhitters.R
\name{funHeavy}
\alias{funHeavy}
\title{Heavyhitters function}
\usage{
funHeavy(x)
}
\arguments{
\item{x}{Vector of categorical values}
}
\value{
Sorted histogram with counts for each level
}
\description{
Function to evaluate most common categorical values
}
