% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism-generalized-laplace.R
\docType{class}
\name{GenLaplace-class}
\alias{GenLaplace-class}
\alias{mechanismGenLaplace}
\title{Generalized Laplace mechanism}
\description{
Laplace mechanism as applied to a generic object which has a specified
add() function that describes how the noise should be added.
}

