% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getFuncArgs}
\alias{getFuncArgs}
\title{Extract function arguments}
\usage{
getFuncArgs(targetFunc, inputList = NULL, inputObject = NULL)
}
\arguments{
\item{targetFunc}{A character vector containing the name of the function to find the matching inputs of
with arguments that need to be filled by output.}

\item{inputList}{A named list of values}

\item{inputObject}{An object with a predefined getFields() function that takes all attributes of object and returns them as a list}
}
\value{
List of arguments and values needed for specification of \code{target.func}
}
\description{
Utility function to match arguments of a function with input list and/or attributes of input object
}
