% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{getHistogramData}
\alias{getHistogramData}
\title{Get Histogram Data}
\usage{
getHistogramData(mechanism, data, variable, var.type, bins)
}
\arguments{
\item{mechanism}{Character, the mechanism used to perturb histogram bins.}

\item{data}{Dataframe, the data to be analyzed with a histogram}

\item{variable}{Character, the name of the variable of of interest in the data frame.}

\item{var.type}{Character, the variable type.}

\item{bins}{Character or numeric, the available bins or levels of a variable. Character 
for categorical variables, a vector of numbers for numeric variables.}
}
\value{
either the column of data corresponding to the variable, or a factor with the
   column of data and the bins as levels.
}
\description{
Determine the bins of the histogram based on the inputs from the user
}
