% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getRegCoefs}
\alias{getRegCoefs}
\title{Function to perform regression using the covariance matrix via the sweep operator}
\usage{
getRegCoefs(formula, release, n, intercept)
}
\arguments{
\item{formula}{Formula}

\item{release}{Numeric private release of covariance matrix}

\item{n}{Integer indicating number of observations}

\item{intercept}{Logical indicating whether the intercept is included}
}
\description{
Function to perform regression using the covariance matrix via the sweep operator
}
