% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-glm.R
\name{glm.getParameters}
\alias{glm.getParameters}
\title{Privacy parameters for GLM}
\usage{
glm.getParameters(accuracy, n, alpha)
}
\arguments{
\item{accuracy}{Numeric giving the accuracy needed to guarantee.}

\item{n}{Integer specifying the number of observations.}

\item{alpha}{Numeric specifying the statistical significance level.}
}
\value{
The scalar epsilon necessary to guarantee the needed accuracy.
}
\description{
Function to find the epsilon value necessary to meet a desired level of 
accuracy for a GLM release.
}

