% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-glm.R
\name{glm.postSummary}
\alias{glm.postSummary}
\title{Summary statistics for differentially private GLM via the bootstrap}
\usage{
glm.postSummary(release, n, model, alpha)
}
\arguments{
\item{release}{Numeric matrix with differentially private estimates for each bootstrap sample}

\item{n}{Integer indicating number of observations}

\item{model}{Character indicating model form}

\item{alpha}{Numeric proportion of vector to be trimmed, specifically the 
least and greatest \code{alpha / 2} are trimmed}
}
\value{
Data frame summary statistics, including estimates and standard errors
}
\description{
Summary statistics for differentially private GLM via the bootstrap
}
