% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-glm.R
\name{glm.release}
\alias{glm.release}
\title{Release a differentially private vector of parameter estimates for a generalized linear model}
\usage{
glm.release(x, n, epsilon, formula, objective, n.boot = NULL,
  intercept = TRUE)
}
\arguments{
\item{x}{Dataframe with data needed to estimate the model}

\item{n}{Integer indicating number of observations}

\item{epsilon}{Numeric epsilon parameter for differential privacy}

\item{formula}{Model formula}

\item{objective}{Perturbed objective function}

\item{n.boot}{Number of bootstrap samples on which to evaluate the 
private parameter estimates, each at \code{epsilon / n.boot} privacy cost}

\item{intercept}{Logical indicating whether the intercept should be added to the model, default TRUE}
}
\description{
Release a differentially private vector of parameter estimates for a generalized linear model
}
\examples{
set.seed(847523)
n <- 1000
x1 <- rnorm(n)
x2 <- rnorm(n)
x3 <- sample(c('a', 'b', 'c', 'd', 'e'), replace=TRUE, size=n)
z <- -0.3 - 1.6 * x1 + 0.3 * x2
p <- 1 / (1 + exp(-z))
y <- rbinom(n, 1, p)
data <- data.frame(cbind(y, x1, x2))
form <- as.formula('y ~ x1 + x2')
logit.true <- glm(form, data=data, family='binomial')$coef
logit.private <- glm.release(data, nrow(data), epsilon=0.5, 
   formula=form, objective=dp.logit)$release
form2 <- as.formula('y ~ x1 + x2 + x3')  # add a factor variable
logit.private2 <- glm.release(data, nrow(data), epsilon=0.5, 
   formula=form2, objective=dp.logit)$release
}
