% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-glm.R
\docType{data}
\name{glmObjectives}
\alias{glmObjectives}
\title{Objective functions}
\format{An object of class \code{list} of length 4.}
\usage{
glmObjectives
}
\description{
List of objective functions support by \code{dpGLM}
}
\keyword{datasets}
