% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{heavyhitters.postNoteFailure}
\alias{heavyhitters.postNoteFailure}
\title{Function to see failed}
\usage{
heavyhitters.postNoteFailure(failed)
}
\arguments{
\item{failed}{something here}
}
\value{
something here
}
\description{
Function to see failed
}
