% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{heavyhitters.release}
\alias{heavyhitters.release}
\title{Release differentially private most common values}
\usage{
heavyhitters.release(x, var.type, epsilon, delta, n, k, bins)
}
\arguments{
\item{x}{Vector of categorical data (character, factor)}

\item{var.type}{Character string indicating data type}

\item{epsilon}{Epsilon value for differential privacy}

\item{n}{Integer indicating number of observations}

\item{k}{Integer querying the most common \code{k} categories}

\item{bins}{Vector of categories from which the top \code{k} categories are evaluated}
}
\value{
A vector whose values are \code{k} most common categories, or a failure message

Contains functions for privately releasing the top-k heavy hitters, that is,
 the k most common unique values.  For example, k=1 releases the mode.
}
\description{
Release differentially private most common values
}
\examples{
N <- 1000
epsilon <- 0.5
delta <- 1e-7
observed.levels <- bins <- c('a', 'b', 'c', 'd', 'e', 'f', 'g', 'h')
probs <- c(0.40, 0.25, 0.15, 0.10, 0.04, 0.03, 0.02, 0.01)
y <- sample(observed.levels, size=N, prob=probs, replace=TRUE)
release <- heavyhitters.release(y, 'character', epsilon, delta, N, k=3, bins=bins)
}
