% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-heavyhitters.R
\name{heavyhittersGetParameters}
\alias{heavyhittersGetParameters}
\title{Heavyhitters epsilon}
\usage{
heavyhittersGetParameters(gap, delta, alpha = 0.05)
}
\arguments{
\item{gap}{The difference between the \code{k} category and the 
\code{k+1} category.}

\item{delta}{A numeric vector representing the probability of an arbitrary
leakage of information from the data. Should be of length one 
and should be a very small value.}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
The epsilon value necessary to gaurantee the given accuracy.
}
\description{
Get the epsilon value necessary to guarantee a desired level of accuracy of a heavyhitters release
}
\author{
Victor Balcer
}
