% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{histogram.formatRelease}
\alias{histogram.formatRelease}
\title{Format the release of private histogram}
\usage{
histogram.formatRelease(release)
}
\arguments{
\item{release}{Table, the result of \code{fun.hist}}
}
\value{
Data frame
}
\description{
Convert the release from a table to a data frame
}
