% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.getJSON}
\alias{histogram.getJSON}
\title{JSON doc for histogram}
\usage{
histogram.getJSON(output.json = TRUE)
}
\arguments{
\item{output.json}{Should the output be converted to JSON format. Default
to \code{TRUE}.}
}
\value{
JSON doc for histogram function.
}
\description{
Produce a JSON doc for differentially private histograms.
}

