% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.getParameters}
\alias{histogram.getParameters}
\title{Privacy parameters}
\usage{
histogram.getParameters(n.bins, n, accuracy, stability, delta = 2^-30,
  alpha = 0.05, error = 1e-09)
}
\arguments{
\item{n.bins}{Integer indicating number of cells in which to tabulate values}

\item{n}{Integer indicating number of observations}

\item{accuracy}{Numeric}

\item{stability}{something here}

\item{delta}{Numeric delta value for differential privacy, fixed}

\item{alpha}{Numeric statistical significance level, fixed}

\item{error}{Numeric, fixed}
}
\value{
Differential privacy parameter epsilon
}
\description{
Privacy parameters
}
