% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.postHerfindahl}
\alias{histogram.postHerfindahl}
\title{Herfindahl index for categorical types}
\usage{
histogram.postHerfindahl(release, n)
}
\arguments{
\item{release}{Numeric vector with noisy counts for each level}

\item{n}{Integer indicating number of observations}
}
\value{
Herfindahl index
}
\description{
Herfindahl index for categorical types
}
