% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{histogramCategoricalBins}
\alias{histogramCategoricalBins}
\title{Utility function to include NA level for categorical types when vector of bins
does not include all observed levels in the data vector.}
\usage{
histogramCategoricalBins(x, bins)
}
\arguments{
\item{x}{Vector, categorical type}

\item{bins}{Vector, depositor-provided list of levels for which to count values}
}
\description{
Utility function to include NA level for categorical types when vector of bins
does not include all observed levels in the data vector.
}
