% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{histogramGetEpsilon}
\alias{histogramGetEpsilon}
\title{Histogram epsilon}
\usage{
histogramGetEpsilon(mechanism, accuracy, delta = 10^-6, alpha = 0.05,
  sensitivity)
}
\arguments{
\item{mechanism}{A string indicating the mechanism that will be used to construct the histogram}

\item{accuracy}{A numeric vector representing the accuracy needed to
guarantee (percent).}

\item{delta}{The probability of an arbitrary leakage of information from
the data. Should be of length one and should be a very small value.
Default to 10^-6.}

\item{alpha}{A numeric vector of length one specifying the numeric
statistical significance level. Default to 0.05.}
}
\value{
Differential privacy parameter epsilon
}
\description{
Function to find the epsilon value necessary to meet a desired level of accuracy for the
differentially private histogram release.
}
\details{
This calculation is the inverse of the calculation for `histogramGetAccuracy`.
}
\seealso{
\code{\link{histogramGetAccuracy}} for accuracy derivation
}
