% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{histogramGetJSON}
\alias{histogramGetJSON}
\title{JSON doc for histogram}
\usage{
histogramGetJSON(output.json = TRUE)
}
\arguments{
\item{output.json}{Should the output be converted to JSON format. Default
to \code{TRUE}.}
}
\value{
JSON doc for histogram function.
}
\description{
Produce a JSON doc for differentially private histograms.
}
