% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{insertChild}
\alias{insertChild}
\title{Insert a child node directly under a specified parent node. If
parent node has no children, will insert a left child. If there
already is a left child, a right child will be inserted where possible,
or an error is raised.}
\arguments{
\item{parent}{Node object.}
}
\description{
Insert a child node directly under a specified parent node. If
parent node has no children, will insert a left child. If there
already is a left child, a right child will be inserted where possible,
or an error is raised.
}
