% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{insertChildren}
\alias{insertChildren}
\title{Insert two children beneath a specified parent node.}
\arguments{
\item{parent}{Node object.}
}
\description{
Insert two children beneath a specified parent node.
}
