% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{inverseVariance}
\alias{inverseVariance}
\title{Calculates variance of Laplace noise.}
\usage{
inverseVariance(sens, eps)
}
\arguments{
\item{sens}{Sensitivity of function that is to be perturbed}

\item{eps}{Privacy parameter used}
}
\description{
Recall that the variance of the Laplace distribution is 2b^2, where b is the classic
scaling parameter of the Laplace distribution (which is in fact the inverse of what we
in this library call the scaling parameter). Then, in our case, b = eps/sens.
}
