% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isMethodDefined}
\alias{isMethodDefined}
\title{Check if a method is defined for a given object.}
\usage{
isMethodDefined(object, methodName)
}
\arguments{
\item{object}{Reference class object.}

\item{methodName}{String. Name of method sought.}
}
\value{
Boolean. TRUE if methodName is a method associated with the object, FALSE otherwise.
}
\description{
Check if a method is defined for a given object.
}
\examples{
Foo <- setRefClass(
   Class='Foo',
   fields = list(
       index = 'numeric'
   ),
   methods = list(
       initialize = function(){
           .self$bar = 1
       }
   )
)

foo <- Foo$new()
isMethodDefined(foo, 'initialize')     # returns TRUE
isMethodDefined(foo, 'add')            # returns FALSE
isMethodDefined(5, 'add')              # returns FALSE

}
