% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{isVector}
\alias{isVector}
\title{Checks if input value is an atomic vector (rather than a list).}
\usage{
isVector(x)
}
\arguments{
\item{x}{arbitrary input}
}
\value{
TRUE if x is an atomic vector, FALSE otherwise.
}
\description{
Checks if input value is an atomic vector (rather than a list).
}
