% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{laplace.getEpsilon}
\alias{laplace.getEpsilon}
\title{Get epsilon for Laplace statistics}
\usage{
laplace.getEpsilon(sensitivity, accuracy, alpha = 0.05)
}
\arguments{
\item{sensitivity}{the sensitivity of the statistic}

\item{accuracy}{A numeric vector representing the accuracy needed to 
guarantee (percent).}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
The scalar epsilon necessary to guarantee the needed accuracy.
}
\description{
Function to find the epsilon value necessary to meet a desired level of 
accuracy for a statistic release.
}
