% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-covariance.R
\name{linearReg}
\alias{linearReg}
\title{Linear regression on covariance matrix}
\usage{
linearReg(formula, release, n, intercept)
}
\arguments{
\item{formula}{An object of class 'formula' containing the desired 
regression formula.}

\item{release}{A numeric private release of the covariance matrix.}

\item{n}{A numeric vector of length one specifying the number of
observations in in the data.}

\item{intercept}{A logical vector indicating whether the intercept is 
included in \code{formula}.}
}
\value{
A numeric vector of regression coefficients corresponding 
   to \code{formula}.
}
\description{
Function to extract regression coefficients from the covariance matrix via 
   the sweep operator.
}
