% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-covariance.R
\name{lowerTriangleSize}
\alias{lowerTriangleSize}
\title{Size of the lower triangle of the covariance matrix that will be generated by covariance statistic.}
\usage{
lowerTriangleSize(columns)
}
\arguments{
\item{columns}{Vector of column names, indicating columns of a data frame 
that the covariance matrix will be calculated over.}
}
\value{
Integer. Size of lower triangle of the covariance matrix that will be generated by covariance statistic. INCLUDING DIAGONAL
}
\description{
Note that covariance matrices are symmetric and thus only the lower triangle needs to be calculated,
so this is equivalent to the number of values that will need to be differentially privately calculated.
}
\details{
If the user input %n\% columns to the dpCovariance initialization, then the
output covariance matrix will have dimension %n \times n \% and thus the 
lower triangle of the matrix will have size %n(n+1)/2\%.
}
