% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_logical}
\alias{make_logical}
\title{Logical variable check}
\usage{
make_logical(x)
}
\arguments{
\item{x}{Vector containing two unique types of values.}
}
\value{
Logical form of \code{x} coded 0-1.

This function effectively allows the user to ask for any variable containing
at most two unique values to treat the variable as logical. If the variable
contains numeric values, the highest value is recoded 1 and the the lower
value is recoded 0. If the variable is categorical and contains only two unique
values, the least frequently observed is recoded 1.
}
\description{
Utility function to verify that a variable is dichotomous.
}
\examples{

make_logical(sample(c('cat', 'dog'), size=8, replace=TRUE))
make_logical(sample(c(0, 1), size=8, replace=TRUE))
make_logical(sample(c(-6.87, 3.23), size=8, replace=TRUE))
}
