% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-objective.R
\name{mapMatrixUnit}
\alias{mapMatrixUnit}
\title{Function to map rows of a numeric matrix to the unit ball}
\usage{
mapMatrixUnit(X, p = 1)
}
\arguments{
\item{X}{Numeric matrix}

\item{p}{The p in p-norm}
}
\value{
List that includes the transformed matrix and the value corresponding to the maximum 
     observed p-norm
}
\description{
This function will ensure that the row having the largest p-norm will be located on the unit ball
}
