% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{matrixToList}
\alias{matrixToList}
\title{Helper function that converts a matrix to a list, s.t. each element of the list is a row of the matrix.}
\usage{
matrixToList(m)
}
\arguments{
\item{m}{An arbitrary matrix}
}
\value{
List form of m, where ith element of list is the ith row of the matrix m.
}
\description{
Helper function that converts a matrix to a list, s.t. each element of the list is a row of the matrix.
}
