% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.getCI}
\alias{mean.getCI}
\title{Confidence Interval for Differentially Private Mean}
\usage{
\method{mean}{getCI}(release, epsilon, sensitivity, n, rng, alpha = 0.05)
}
\arguments{
\item{release}{Differentially private release of mean of vector x}

\item{epsilon}{Numeric, epsilon parameter for differential privacy}

\item{sensitivity}{Numeric, the sensitivity of the estimate}

\item{n}{The number of observations in the data}

\item{rng}{An a priori estimate of the range}

\item{alpha}{The statistical signifcance level}
}
\value{
Confidence bounds for differentially private release
}
\description{
Confidence Interval for Differentially Private Mean
}
