% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.postHistogram}
\alias{mean.postHistogram}
\title{Postprocessed histogram for logical variables}
\usage{
\method{mean}{postHistogram}(release, n)
}
\arguments{
\item{release}{Numeric private mean}

\item{n}{Integer indicating number of observations}
}
\value{
Data frame, histogram of the logical variable
}
\description{
Generate counts for levels of a logical variable based on the release
}
