% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.postMedian}
\alias{mean.postMedian}
\title{Postprocessed Median for Logical Variables}
\usage{
\method{mean}{postMedian}(release)
}
\arguments{
\item{release}{Differentially private release of a mean for a logical 
variable.}
}
\value{
Median of \code{release}.
}
\description{
Postprocessed Median for Logical Variables
}
