% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.postStandard_Deviation}
\alias{mean.postStandard_Deviation}
\title{Postprocessed Standard Deviation for Logical Variables}
\usage{
\method{mean}{postStandard_Deviation}(release)
}
\arguments{
\item{release}{Differentially private release of a mean for a logical variable}
}
\description{
Postprocessed Standard Deviation for Logical Variables
}
