% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mean.R
\name{mean.postStandardDeviation}
\alias{mean.postStandardDeviation}
\title{Postprocessed standard deviation for logical variables}
\usage{
\method{mean}{postStandardDeviation}(release)
}
\arguments{
\item{release}{Differentially private release of a mean for a logical
variable.}
}
\value{
Standard deviation of the logical variable
}
\description{
Calculates the standard deviation of the differentially private mean from a
logical variable.
}
