% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mean.R
\name{meanGetJSON}
\alias{meanGetJSON}
\title{JSON doc for mean}
\usage{
meanGetJSON(outputJSON = TRUE)
}
\arguments{
\item{outputJSON}{Should the output be converted to JSON format. Default
to \code{TRUE}.}
}
\value{
JSON for mean function
}
\description{
Produce a JSON doc for differentially private means.
}
