% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mean.R
\name{meanPostMedian}
\alias{meanPostMedian}
\title{Postprocessed median for logical variables}
\usage{
meanPostMedian(release)
}
\arguments{
\item{release}{Differentially private release of a mean for a logical
variable.}
}
\value{
Median of the logical variable
}
\description{
Calculates the median of the differentially private mean from a
logical variable.
}
