% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mean.R
\name{meanSensitivity}
\alias{meanSensitivity}
\title{Sensitivity of mean}
\usage{
meanSensitivity(rng, n)
}
\arguments{
\item{rng}{Numeric vector of length two; first entry is minimal bound on the database entries, second is maximal bound on the database entries.}

\item{n}{Numeric vector of length one; the number of datapoints in the database.}
}
\value{
Numeric vector of length one; a maximal bound on the sensitivity of the population variance.
}
\description{
For a detailed derivation of the sensitivity, see /extra_docs/mean_sensitivity.pdf
}
\examples{
meanSensitivity(c(0,10),5) #should return 2
}
