% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{mechanism.exponential}
\alias{mechanism.exponential}
\title{Exponential mechanism}
\usage{
mechanism.exponential(fun, x, var.type, sensitivity, epsilon, k,
  postlist = NULL, ...)
}
\description{
Exponential mechanism
}
