% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{mechanism.gaussian}
\alias{mechanism.gaussian}
\title{Gaussian mechanism}
\usage{
mechanism.gaussian(fun, x, var.type, rng, sensitivity, epsilon, delta,
  postlist = NULL, ...)
}
\description{
Gaussian mechanism
}
