% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{mechanism.histogram.random}
\alias{mechanism.histogram.random}
\title{Describe Here}
\usage{
mechanism.histogram.random(x, var_type, epsilon, levels, bins, n_bins)
}
\arguments{
\item{x}{something here}

\item{var_type}{something here}

\item{epsilon}{something here}

\item{levels}{something here}

\item{bins}{something here}

\item{n_bins}{something here}
}
\value{
something here
}
\description{
Describe Here
}

