% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{mechanism.objective}
\alias{mechanism.objective}
\title{Objective perturbation}
\usage{
mechanism.objective(fun, x, n, epsilon, n.boot, postlist, ...)
}
\arguments{
\item{fun}{A user supplied function or string vector of length one naming a 
function.}

\item{x}{A data frame of variables to run \code{fun} on.}

\item{n}{Integer indicating number of observations.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{n.boot}{An integer indicating the number of bootstrap estimations
the user would like to perform.}

\item{postlist}{A list with names, function pairs for post-processing 
statistics.}

\item{...}{Other arguments.}
}
\description{
Mechanism to add noise from a LaPlace distribution to the objective function of the 
   specified regression model for differentially private regression.
}
