% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mpinv}
\alias{mpinv}
\title{Moore Penrose Inverse Function ###Must assign authorship to this###}
\usage{
mpinv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{A numeric, symmetric covariance matrix.}

\item{tol}{Convergence requirement.}
}
\description{
Generate the Moore-Penrose pseudoinverse matrix of \code{X}.
}
