% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\alias{Node}
\title{Node class}
\value{
Node object.
}
\description{
Node class
}
\section{Fields}{

\describe{
\item{\code{index}}{numeric. Index of node in tree structure. Tree is indexed from top to bottom, left to right.}

\item{\code{parent}}{Node object. Points to parent node of given node.}

\item{\code{leftChild}}{Node object. Points to left child of given node.}

\item{\code{rightChild}}{Node object. Points to right child of given node.}

\item{\code{weight}}{numeric. Numeric weight of a given node. E.g. if tree used to
bin points into set ranges, how many points are in the range spanned by that node.}

\item{\code{depth}}{numeric. Numeric depth of the node within tree structure.}

\item{\code{range}}{ANY. Numeric array c(min,max) indicating range represented by the node, where min is minimum of range and
max is maximum of range.}
}}


