% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{normalizeHistogram}
\alias{normalizeHistogram}
\title{Constrain the sum of histogram bins to sample size}
\usage{
normalizeHistogram(h, n)
}
\arguments{
\item{h}{Histogram}

\item{n}{Sample size}
}
\value{
The noisy histogram with the sum of the bin counts normalized to the input data size
}
\description{
Constrain the sum of histogram bins to sample size
}
