% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{normalizeReleaseAndConvertToDataFrame}
\alias{normalizeReleaseAndConvertToDataFrame}
\title{Convert normalized histogram release to data frame}
\usage{
normalizeReleaseAndConvertToDataFrame(release, n)
}
\arguments{
\item{release}{Table, the result of \code{funHist}}

\item{n}{Sample size}
}
\value{
Data frame
}
\description{
Convert the release of the private histogram from a table to a data frame.
While converting, the normalizeHistogram method is called, which normalizes
the private histogram so that the sum of the bin counts is equal to n.
}
