% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{optimalCount}
\alias{optimalCount}
\title{Optimal counts for nodes of tree}
\usage{
optimalCount(tree, wA, cA, wB, cB)
}
\arguments{
\item{tree}{Tree, formatted as a list of arrays where the contents of the ith array in the list is the ith level of the tree.}

\item{wA}{Array of weights calculated from above with wAbove}

\item{cA}{Array of counts calculated from below with countBelow}

\item{wB}{Array of weights calculated from below with wBelow}

\item{cB}{Array of counts calculated from above with countAbove}
}
\value{
List of counts for each node of the tree.
}
\description{
See extra_docs/tree-post-processing for formula. This assumes that the variance is the same for every node at in the tree.
}
\examples{
t <- list(c(10), c(6,4), c(3,3,1,3))
wB <- wBelow(t)
wA <- wAbove(t, wB)
cB <- countBelow(t, wB)
cA <- countAbove(t, cB, wA)
c <- optimalCount(t, wA, cA, wB, cB) #will return t

}
