% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{optimalSigma}
\alias{optimalSigma}
\title{Optimal Sigma Estimation}
\usage{
optimalSigma(wA, wB, epsilon)
}
\arguments{
\item{wA}{Array of weights calculated from above}

\item{wB}{Array of weights calculated from below}

\item{epsilon}{Epsilon used for Laplace noise addition}
}
\value{
Value of variance for optimal estimate
}
\description{
See extra_docs/tree-post-processing for formula. This assumes that the variance is the same for every node at in the tree.
}
