% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plap}
\alias{plap}
\title{Cumulative distribution function for Laplace distribution}
\usage{
plap(x, mu = 0, b = 1)
}
\arguments{
\item{x}{numeric, value}

\item{mu}{numeric, center of the distribution}

\item{b}{numeric, spread}
}
\value{
Probability less than or equal to x
}
\description{
Cumulative distribution function for Laplace distribution
}
\examples{
x <- 0
plap(x)
}

