% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{printTree}
\alias{printTree}
\title{Simple print function to print tree structure for debugging.. Recurses through the nodes of the tree and prints 
information associated with each node.}
\arguments{
\item{currNode}{Node object, defaults to the head of the tree.}
}
\description{
Simple print function to print tree structure for debugging.. Recurses through the nodes of the tree and prints 
information associated with each node.
}
