% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PUMS5extract10000}
\alias{PUMS5extract10000}
\title{Census California Public Use Micro Sample (PUMS) Dataset}
\format{A data set with 10000 observations and 11 variables}
\source{
From Honaker, James, 2015, "PUMS California Subsample",
doi:10.5072/FK2/PXQ0KE, Beta Dataverse, V2
}
\description{
Census California Public Use Micro Sample (PUMS) Dataset
}
\keyword{datasets}
