% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{qlap}
\alias{qlap}
\title{Quantile function for Laplace distribution}
\usage{
qlap(p, mu = 0, b = 1)
}
\arguments{
\item{p}{Numeric, vector of probabilities}

\item{mu}{numeric, center of the distribution}

\item{b}{numeric, spread}
}
\value{
Quantile function
}
\description{
Quantile function for Laplace distribution
}
\examples{
probs <- c(0.05, 0.50, 0.95)
qlap(probs)
}

