% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-quantiles.R
\name{quantile.computeInterval}
\alias{quantile.computeInterval}
\title{Utility function to compute the count in a range from a binary tree}
\usage{
\method{quantile}{computeInterval}(tree, val, range, gran)
}
\arguments{
\item{tree}{The binary tree stored as an array}

\item{val}{The value we want to compute the count in [min, val] for}

\item{range}{The range of the universe as a vector (min, max)}

\item{gran}{The granularity of the universe between the min and max}
}
\value{
The count in [min, val] given by the binary tree.
}
\description{
Utility function to compute the count in a range from a binary tree
}
\author{
Nathan Manohar
}

