% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-quantiles.R
\name{quantile.getParameters}
\alias{quantile.getParameters}
\title{Get the epsilon value necessary to guarantee a desired level of accuracy of a quantile release}
\usage{
\method{quantile}{getParameters}(alpha, beta, range, gran, n)
}
\arguments{
\item{alpha}{the accuracy parameter}

\item{beta}{the true value is within the accuracy range (alpha)}

\item{range}{The range of the universe as a vector (min, max)}

\item{gran}{The granularity of the universe between the min and max}

\item{n}{Number of observations of the variable}
}
\value{
The epsilon value necessary to gaurantee the given accuracy
}
\description{
Get the epsilon value necessary to guarantee a desired level of accuracy of a quantile release
}
\author{
Nathan Manohar
}

