% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-quantiles.R
\name{quantile.release}
\alias{quantile.release}
\title{Release differentially private quantiles}
\usage{
\method{quantile}{release}(x, var.type, n, epsilon, rng, gran, cdf.step)
}
\description{
Release differentially private quantiles
}

