% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{regress}
\alias{regress}
\title{Function to perform regression using the covariance matrix via the sweep operator}
\usage{
regress(formula, release, n, intercept)
}
\arguments{
\item{formula}{Formula}

\item{release}{Numeric private release of covariance matrix}

\item{n}{Integer indicating number of observations}

\item{intercept}{Logical indicating whether the intercept is included}
}
\value{
something here
}
\description{
Function to perform regression using the covariance matrix via the sweep operator
}
