% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-regression.R
\name{regression.release}
\alias{regression.release}
\title{Function to perform regression using the differentially private covariance matrix via the sweep operator}
\usage{
regression.release(formula, release, n)
}
\arguments{
\item{formula}{Formula}

\item{release}{Numeric private release of covariance matrix}

\item{n}{Integer indicating number of observations}

\item{intercept}{Logical indicating whether the intercept is included}
}
\description{
Function to perform regression using the differentially private covariance matrix via the sweep operator
}
