% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-postprocessing.R
\name{release2json}
\alias{release2json}
\title{Create json file of metadata from list of release objects}
\usage{
release2json(release, nameslist)
}
\arguments{
\item{release}{a list of release objects}
}
\description{
Create json file of metadata from list of release objects
}
