% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rlaplace}
\alias{rlaplace}
\title{Random draw from Laplace distribution}
\usage{
rlaplace(n = 1, sensitivity, epsilon)
}
\arguments{
\item{n}{integer, number of draws}

\item{sensitivity}{numeric}

\item{epsilon}{numeric}
}
\value{
Random draws from Laplace distribution
}
\description{
Random draw from Laplace distribution
}
\examples{
rlaplace(sensitivity=1, epsilon=0.1)
}
