% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{setHistogramRange}
\alias{setHistogramRange}
\title{Set histogram range if bins are entered}
\usage{
setHistogramRange(rng, varType, bins)
}
\arguments{
\item{rng}{The rng entered by the user, may be NULL}

\item{bins}{The histogram bins entered by the user, may be NULL}

\item{delta}{The variable type of the data}
}
\value{
The 2-tuple that will be used as the range for the histogram when 
calculating sensitivity, censoring data, and imputing values.
}
\description{
If the user enters a logical variable, they should not need to enter a
histogram range, but calculating the sensitivity requires a range. This
method sets the range for a logical variable as c(0,1).
}
\details{
If the user enters numeric bins, they should not need to enter a histogram
range, but calculating the sensitivity requires a range. This method sets
the range for a numeric variable with bins entered to the bins lower bound
and the bin upper bound.

If neither of the cases are true for the histogram, the range the user entered
if returned.
}
