% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{setNumHistogramBins}
\alias{setNumHistogramBins}
\title{Set the number of histogram bins}
\usage{
setNumHistogramBins(n.bins, granularity, var.type, bins)
}
\arguments{
\item{n.bins}{the number of bins entered by the user, may be null}

\item{granularity}{the number of items to be in each bin (i.e. the height of each bin), may be null}

\item{var.type}{The variable type of the data that was entered by the user}

\item{bins}{the bin vector either entered by the user or set by determineBins()}
}
\value{
No return value, will only send an error message if variable tyep is invalid.
}
\description{
Set the number of bins for the histogram, given either the entered number of bins,
the entered granularity, or the vector of bins.
}
