% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{setVariableTypeAsFactor}
\alias{setVariableTypeAsFactor}
\title{Set Variable Type as Factor}
\usage{
setVariableTypeAsFactor(object)
}
\arguments{
\item{object}{The dpHistogram object for the histogram on the specific variable. The object
has a 'varType' member that will be changed from 'logical' to 'factor'}
}
\value{
No return value
}
\description{
Set the variable type of the input variable to type 'factor'. This will only be used
when the input variable type of 'logical' and impute == FALSE.
}
