% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sgn}
\alias{sgn}
\title{Sign function}
\usage{
sgn(x)
}
\arguments{
\item{x}{numeric, value or vector or values}
}
\value{
The sign of passed values
}
\description{
Sign function
}
\examples{
sgn(rnorm(10))
}
