% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism-snapping.R
\name{Snapping Mechanism}
\alias{Snapping Mechanism}
\title{Snapping Mechanism}
\arguments{
\item{fun}{function of input x for which we will add noise to the output.}

\item{x}{input that function fun will be evaluated on.}

\item{sens}{sensitivity of fun. Sensitivity is defined in above citation.}

\item{postFun}{post-processing function. Takes differentially private release as input
and returns some form of output in principal based on the differentially private release.}

\item{...}{any additional (optional) parameters}
}
\value{
result of post-processing on input function "fun" evaluated on database "x", assuming sensitivity of fun is "sens".
}
\description{
Differentially private evaluation of input function "fun" with sensitivity "sens" on input data
"x" using the Snapping mechanism.
}
\references{
Ilya Mironov,  On Significance of the Least Significant Bits for Differential Privacy. 2012
}
