% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-noise-generation.R
\name{snapping.getAccuracy}
\alias{snapping.getAccuracy}
\title{Get accuracy for Snapping statistics}
\usage{
snapping.getAccuracy(sensitivity, epsilon, alpha = 0.05)
}
\arguments{
\item{sensitivity}{the sensitivity of the statistic}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
Accuracy guarantee for statistic release given epsilon.
}
\description{
Function to find the accuracy guarantee of a statistic release at a given epsilon
value.
}
