% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-noise-generation.R
\name{snapping.getEpsilon}
\alias{snapping.getEpsilon}
\title{Get epsilon for Snapping statistics}
\usage{
snapping.getEpsilon(sensitivity, accuracy, alpha = 0.05)
}
\arguments{
\item{sensitivity}{the sensitivity of the statistic}

\item{accuracy}{A numeric vector representing the accuracy needed to
guarantee (percent).}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
The scalar epsilon necessary to guarantee the needed accuracy.
}
\description{
Function to find the epsilon value necessary to meet a desired level of
accuracy for a statistic release.
}
