% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism-stability.R
\name{Stability Mechanism}
\alias{Stability Mechanism}
\title{Stability Mechanism}
\arguments{
\item{fun}{function of input x to add Laplace noise to.}

\item{x}{input that function fun will be evaluated on.}

\item{sens}{sensitivity of fun. Sensitivity is defined in above citation.}

\item{postFun}{post-processing function. Takes differentially private release as input
and returns some form of output in principal based on the differentially private release.}

\item{...}{any additional (optional) parameters}
}
\value{
result of post-processing on input function "fun" evaluated on database "x", assuming sensitivity of fun is "sens".
}
\description{
Differentially private evaluation of input function "fun" with sensitivity "sens" on input data 
"x" using the Stability mechanism. The Stability mechanism is meant to be used only with stable
functions, which have a stronger guarantee on the output between neighboring databases. See 
References for more information on stable functions.
}
\details{
The Stability mechanism should only be used on the histogram statistic in the PSIlence library.
For a histogram generated by the Stability mechanism, empty buckets will be removed and any
buckets with a count below an accuracy threshold will also be removed. Removing these buckets based on
low counts is what creates the added guarantee of privacy.

The accuracy threshold for removing bins with low counts is calculated by: \eqn{1 + (2 * log(2 / \delta) / \epsilon)}.
}
\examples{
# the function in `statistic-histogram.R` that creates a histogram from input data
histogram_function <- fun.hist 
# the data frame that holds the data we want to analyze, in this case the data is called "PUMS5extract10000"
data <- data(PUMS5extract10000) 
# the variable for which we want a histogram
variable <- "age"
# the sensitivity for the histogram, the default sensitivity for histograms is 2 
sens <- 2 
# the post-processing function to use to format the histogram release correctly
post_processing_function <- dpHistogram$postProcess 

stability_histogram <- mechanismStability$evaluate(histogram_function, data[, variable], sens, post_processing_function)

}
\references{
S. Vadhan The Complexity of Differential Privacy, Section 3.3 Releasing Stable Values p.23-24. March 2017.
}
