% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{stabilityGetAccuracy}
\alias{stabilityGetAccuracy}
\title{Get accuracy for a stable statistic (only used for histogram statistic)}
\usage{
stabilityGetAccuracy(sensitivity, epsilon, delta = 2^-30, alpha = 0.05)
}
\arguments{
\item{sensitivity}{the sensitivity of the statistic}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{delta}{The probability of an arbitrary leakage of information from 
the data. Should be of length one and should be a very small value. 
Default to 2^-30.}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
Accuracy guarantee for statistic release given epsilon.
}
\description{
Function to find the accuracy guarantee of a stable statistic release at a given epsilon 
value.
}
