% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-histogram.R
\name{stabilityGetEpsilon}
\alias{stabilityGetEpsilon}
\title{Get epsilon for a stable statistic (only used for histogram statistic)}
\usage{
stabilityGetEpsilon(sensitivity, accuracy, delta = 2^-30, alpha = 0.05)
}
\arguments{
\item{sensitivity}{the sensitivity of the statistic}

\item{accuracy}{A numeric vector representing the accuracy needed to 
guarantee (percent).}

\item{delta}{The probability of an arbitrary leakage of information from 
the data. Should be of length one and should be a very small value. 
Default to 2^-30.}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
The scalar epsilon necessary to guarantee the needed accuracy.
}
\description{
Function to find the epsilon value necessary to meet a desired level of 
accuracy for a stable statistic release.
}
