% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stErr}
\alias{stErr}
\title{Function to evaluate the standard error of a node estimate given a weight and the standard
 deviation of the noise used to perturb the nodes}
\usage{
stErr(w, sigma)
}
\arguments{
\item{w}{Weight used construct the estimate}

\item{sigma}{Standard deviation of the noise used to perturb the estimates}
}
\value{
Standard error of the node estimate
}
\description{
Function to evaluate the standard error of a node estimate given a weight and the standard
 deviation of the noise used to perturb the nodes
}

