% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{traverseLeft}
\alias{traverseLeft}
\title{Determine whether or not the targetNode is to the left of the
current node (currNode) in the tree, assuming the targetNode is beneath the
current node.}
\arguments{
\item{currNodeIndex}{Positive integer, index of current node in tree.}

\item{targetNodeIndex}{Positive integer, index of target node in tree.}

\item{x}{Numeric. Data point to be binned.}

\item{currNode}{Node object. The current node to bin point into.}
}
\value{
Boolean. TRUE if targetNode is to the left of currNode, FALSE otherwise.

Boolean. TRUE if should traverse left, FALSE else.

E.g. given currNode node with range [0,4), its left child will have range [0,2) and its
right child will have range [2,4). Then data point x = 1 should be binned into the left child so 
traverseLeft(1,currNode) will return TRUE. 

If instead x = 2, it should be binned into the right child, so traverseLeft(2,currNode) would 
return FALSE.
}
\description{
At node currNode, if x is in the range of the left child, should traverse left
down tree and increment weight of the left child. Otherwise, should traverse right.
}
