% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-tree.R
\name{tree.getAccuracy}
\alias{tree.getAccuracy}
\title{Accuracy for a differentially private binary tree}
\usage{
tree.getAccuracy(epsilon, rng, gran, alpha = 0.05)
}
\arguments{
\item{epsilon}{Numeric differential privacy parameter}

\item{rng}{Numeric a priori estimate of the variable range}

\item{gran}{Numeric granularity}

\item{alpha}{Numeric level of statistical significance, default 0.05}
}
\value{
Accuracy guarantee for the tree given epsilon
}
\description{
Accuracy for a differentially private binary tree
}
