% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\name{tree.postCDF}
\alias{tree.postCDF}
\title{Function to derive CDF from efficient terminal node counts}
\usage{
tree.postCDF(release, rng, terminal.index)
}
\arguments{
\item{release}{Efficient differentially private binary tree}

\item{rng}{An a priori estimate of the range of the vector
being represented as a binary tree}

\item{terminal.index}{Vector of indices corresponding to the terminal
leaf nodes of the binary tree}
}
\value{
Differentially private estimate of the empirical cumulative
     distribution function
}
\description{
Function to derive CDF from efficient terminal node counts
}
