% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\name{tree.postEfficient}
\alias{tree.postEfficient}
\title{Function to efficiently estimate noisy node counts}
\usage{
tree.postEfficient(release, tree.data, n, variance, terminal.index)
}
\arguments{
\item{release}{The truncated differentially private noisy binary tree
in vector form}

\item{tree.data}{Data frame with binary tree attributes, including depth
and indicators of parent and adjacent nodes. Note that
\code{nrow(tree.data) == length(release)}}

\item{n}{Number of observations}

\item{variance}{The variance of the noise used to perturb tree nodes}

\item{terminal.index}{Vector of indices corresponding to the terminal
leaf nodes of the binary tree}

\item{n.nodes}{Number of nodes in the binary tree, also \code{length(release)}}
}
\value{
Efficient differentially private binary tree
}
\description{
Function to efficiently estimate noisy node counts
}
