% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-tree.R
\name{tree.postEfficientTree}
\alias{tree.postEfficientTree}
\title{Function to efficiently estimate noisy node counts}
\usage{
tree.postEfficientTree(release, tree.data, n, n.nodes, sigma, inv.sigma.sq,
  terminal.index)
}
\arguments{
\item{release}{The truncated differentially private noisy binary tree
in vector form}

\item{tree.data}{Data frame with binary tree attributes, including depth
and indicators of parent and adjacent nodes. Note that
\code{nrow(tree.data) == length(release)}}

\item{n}{Number of observations}

\item{n.nodes}{Number of nodes in the binary tree, also \code{length(release)}}

\item{sigma}{The standard deviation of the noise used in perturbing nodes}

\item{inv.sigma.sq}{Inverse variance of the noise used in perturbing nodes}

\item{terminal.index}{Vector of indices corresponding to the terminal
leaf nodes of the binary tree}
}
\value{
Efficient differentially private binary tree
}
\description{
Function to efficiently estimate noisy node counts
}
