% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\name{tree.postMedian}
\alias{tree.postMedian}
\title{Function to evaluate the median using the DP CDF}
\usage{
tree.postMedian(cdf)
}
\arguments{
\item{cdf}{Differentially private estimate of the empirical cumulative
distribution function}
}
\value{
Differentially private estimate of the median
}
\description{
Function to evaluate the median using the DP CDF
}
