% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-tree.R
\name{tree.release}
\alias{tree.release}
\title{Function to release a differentially private tree and post-processing}
\usage{
tree.release(x, var.type, n, epsilon, rng, gran, percentiles = NULL)
}
\arguments{
\item{x}{Vector of numeric observations}

\item{var.type}{Character vector specifying the variable type, should be
one of \code{c('numeric', 'integer')}}

\item{n}{Number of observations}

\item{epsilon}{Privacy parameter epsilon, should be between zero and one}

\item{rng}{An a priori estimate of the range of \code{x}}

\item{gran}{The granularity at which \code{x} is represented in the tree}

\item{percentiles}{Vector of percentiles used in the post-processing
quantile function. The default is \code{NULL}, in which case the
quantile function is not executed.}
}
\value{
List of values including the differentially private tree, the
     cumulative distribution function, the median, and optionally a
     vector of percentiles. Other attributes of the binary tree are
     also included.
}
\description{
Function to release a differentially private tree and post-processing
}
