% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\name{treeBins}
\alias{treeBins}
\title{Calculate bins of each leaf in the tree by row}
\usage{
treeBins(rng, depth, n)
}
\arguments{
\item{rng}{Total range of data to be binned into tree in form (min, max) A tuple of numerics of length 2.}

\item{depth}{Depth of the tree, considering the root to have depth 0.}

\item{n}{Number of data points to be binned.

Note: This can be calculated entirely publically. The fact that 'n' is input is unnecessary,
but is a residual affect of the fact that the helper function called can also be used if the user
specifies their desired granularity of the bins instead of supplying a range, which may be something 
we want to eventually implement here.}
}
\value{
A list of the bins used in the tree, from the first level of the tree to the leaves.
}
\description{
Calculate bins of each leaf in the tree by row
}
