% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\name{treeGetParameters}
\alias{treeGetParameters}
\title{Epsilon for a differentially private binary tree}
\usage{
treeGetParameters(accuracy, rng, gran, alpha = 0.05)
}
\arguments{
\item{accuracy}{Numeric accuracy needed}

\item{rng}{Numeric a priori estimate of the variable range}

\item{gran}{Numeric granularity}

\item{alpha}{Numeric level of statistical significance, default 0.05}
}
\value{
Epsilon necessary to guarantee the given accuracy
}
\description{
Epsilon for a differentially private binary tree
}
