% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{treeMean}
\alias{treeMean}
\title{Mean of Input Data}
\usage{
treeMean(tree, bins)
}
\arguments{
\item{tree}{Differentially private tree of histogram counts generated with dpTree.}

\item{bins}{The bins by level of the tree, which corresponds to the binsByLevel attribute of dpTree.}
}
\value{
Estimate of the mean of the underlying data that created the dpTree.
}
\description{
Estimated mean of input data.
}
\details{
Given a histogram of numeric values, you can estimate the mean of the underlying data by calculating the midpoint of each of the bins.

Let c_i be the count associated with bin i, and let mid_i be the midpoint of the data range that bin i represents.
Let n be the total number of data points. Then, 
\deqn{(1/n)\sum c_i * mid_i}  
is an estimate of the mean.

Here, we calculate this estimate using the highest granularity bins possible, so the histogram created by the leaf nodes of the tree.
}
