% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\name{treePostEfficient}
\alias{treePostEfficient}
\title{Function to efficiently estimate noisy node counts}
\usage{
treePostEfficient(release, treeData, n, variance, terminalIndex)
}
\arguments{
\item{release}{The truncated differentially private noisy binary tree
in vector form}

\item{treeData}{Data frame with binary tree attributes, including depth
and indicators of parent and adjacent nodes. Note that
\code{nrow(treeData) == length(release)}}

\item{n}{Number of observations}

\item{variance}{The variance of the noise used to perturb tree nodes}

\item{terminalIndex}{Vector of indices corresponding to the terminal
leaf nodes of the binary tree}

\item{nNodes}{Number of nodes in the binary tree, also \code{length(release)}}
}
\value{
Efficient differentially private binary tree
}
\description{
Function to efficiently estimate noisy node counts
}
