% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\name{treePostFormatRelease}
\alias{treePostFormatRelease}
\title{Function to truncate negative noisy node counts at zero}
\usage{
treePostFormatRelease(release)
}
\arguments{
\item{release}{The differentially private noisy binary tree}
}
\value{
Noisy binary tree truncated at zero
}
\description{
Function to truncate negative noisy node counts at zero
}
