% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-tree.R
\name{treePostMean}
\alias{treePostMean}
\title{Function to evaluate the mean using the DP CDF}
\usage{
treePostMean(cdf, rng)
}
\arguments{
\item{cdf}{Differentially private estimate of the empirical cumulative
distribution function}

\item{rng}{Numeric a priori estimate of the range}

\item{gran}{Granularity}
}
\value{
Differentially private estimate of the mean
}
\description{
Function to evaluate the mean using the DP CDF
}
