% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{treePostMedian}
\alias{treePostMedian}
\title{Function to evaluate the median using the DP CDF}
\usage{
treePostMedian(cdf)
}
\arguments{
\item{cdf}{Differentially private estimate of the empirical cumulative
distribution function}
}
\value{
Differentially private estimate of the median
FIX THIS
}
\description{
Function to evaluate the median using the DP CDF
}
