% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trimVector}
\alias{trimVector}
\title{Function to trim lower and upper regions of a vector of values}
\usage{
trimVector(vec, alpha)
}
\arguments{
\item{vec}{Numeric vector}

\item{alpha}{Numeric proportion of vector to be trimmed, specifically the 
least and greatest \code{alpha / 2} are trimmed}
}
\value{
Trimmed vector
}
\description{
Function to trim lower and upper regions of a vector of values
}
