% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-variance.R
\name{variance.post.std}
\alias{variance.post.std}
\title{Function to extract standard deviation from noisy estimate of variance}
\usage{
variance.post.std(release)
}
\arguments{
\item{release}{Numeric noisy estimate of variance}
}
\value{
Noisy estimate of standard deviation
}
\description{
Function to extract standard deviation from noisy estimate of variance
}
