% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-variance.R
\name{variance.postStandardDeviation}
\alias{variance.postStandardDeviation}
\title{Postprocessed variance standard deviation}
\usage{
variance.postStandardDeviation(release)
}
\arguments{
\item{release}{Differentially private release of variance.}
}
\value{
Noisy estimate of the standard deviation of \code{release}.
}
\description{
Function to extract standard deviation from noisy estimate of variance.
}

