% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-variance.R
\name{variance.release}
\alias{variance.release}
\title{Release differentially private variance}
\usage{
variance.release(x, var.type, n, epsilon, rng)
}
\arguments{
\item{x}{A numeric, logical, or integer vector.}

\item{var.type}{A character vector specifying variable type of \code{x}. 
Should be of length one and should contain either 'numeric', 
'logical', or 'integer'.}

\item{n}{A numeric vector of length one specifying the number of
observations in \code{x}.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{rng}{A numeric vector specifying an a priori estimate of the range
of \code{x}. Should be of length two.}
}
\value{
A list output containing noisy estimate and post-processing values from the 
   Laplace mechanism.
}
\description{
Function to calculate a differentially private release of variance.
}
\examples{

}
