% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-variance.R
\name{variancePostStandardDeviation}
\alias{variancePostStandardDeviation}
\title{Postprocessed variance standard deviation}
\usage{
variancePostStandardDeviation(release)
}
\arguments{
\item{release}{Differentially private release of variance.}
}
\value{
Noisy estimate of the standard deviation of \code{release}.
}
\description{
Function to extract standard deviation from noisy estimate of variance.
}
