% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-variance.R
\name{varianceSensitivity}
\alias{varianceSensitivity}
\title{Sensitivity of sample variance}
\usage{
varianceSensitivity(n, rng)
}
\arguments{
\item{n}{Numeric vector of length one; the number of datapoints in the database.}

\item{rng}{Numeric vector of length two; first entry is minimal bound on the database entries, second is maximal bound on the database entries.}
}
\value{
Numeric vector of length one; a maximal bound on the sensitivity of the sample variance.
}
\description{
For a detailed derivation of the sensitivity, see /extra_docs/variance_sensitivity.pdf.
}
\examples{
varianceSensitivity(2,c(0,10)) #should return 50
}
