% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vectorNorm}
\alias{vectorNorm}
\title{Function to evaluate the p-norm of vectors in a matrix}
\usage{
vectorNorm(X, p = 1, margin = 1)
}
\arguments{
\item{X}{matrix of numeric values}

\item{p}{The p in p-norm}

\item{margin}{The subscripts over which the norm is applied, where 1 gives row norms
and 2 gives the column norms}
}
\value{
A vector of norms
}
\description{
Function to evaluate the p-norm of vectors in a matrix
}
