% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wAbove}
\alias{wAbove}
\title{Function to evaluate weights from the noise variance and standard errors in a parent and adjacent 
 nodes for the node of a differentially private binary tree}
\usage{
wAbove(inv.sigma.sq, tree, parent, adjacent)
}
\arguments{
\item{inv.sigma.sq}{Inverse variance of the noise used in perturbing nodes}

\item{tree}{Data frame with binary tree attributes and node values}

\item{parent}{Index of the parnet node}

\item{adjacent}{Index of the adjacent node}
}
\value{
Weight
}
\description{
Function to evaluate weights from the noise variance and standard errors in a parent and adjacent 
 nodes for the node of a differentially private binary tree
}

