% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wEfficient}
\alias{wEfficient}
\title{Function to evaluate weights efficiently using the noise variance and standard errors in parent and adjacent 
 nodes as well child nodes for the node of a differentially private binary tree}
\usage{
wEfficient(tree, idx, parent, adjacent)
}
\arguments{
\item{tree}{Data frame with binary tree attributes and node values}

\item{idx}{Index of the node for which the weight is evaluated}

\item{parent}{Index of the parnet node}

\item{adjacent}{Index of the adjacent node}
}
\value{
Weight
}
\description{
Function to evaluate weights efficiently using the noise variance and standard errors in parent and adjacent 
 nodes as well child nodes for the node of a differentially private binary tree
}

