% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{censordata}
\alias{censordata}
\title{Utility function for censoring data}
\usage{
censordata(x, var_type, range = NULL, levels = NULL)
}
\arguments{
\item{x}{A vector of values to censor within a range}

\item{range}{A vector (min, max) of the bounds of the range}
}
\value{
Original vector with values outide the bounds of the range censored to the bounds

Checks if val is in range = (min, max)
  and censors values to either min or max if it is 
  out of the range
}
\description{
Utility function for censoring data
}
\examples{
censordata(x=1:10, var_type='integer', range=c(2.5,7)) 
}

