% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_histogram_bins}
\alias{check_histogram_bins}
\title{Utility function to check bins argument to histogram}
\usage{
check_histogram_bins(n_bins, n)
}
\arguments{
\item{n_bins}{Number of cells in which to tabulate values}

\item{n}{Number of observations

If number of bins is not provided, use the Sturges method}
}
\description{
Utility function to check bins argument to histogram
}

