% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_histogram_n}
\alias{check_histogram_n}
\title{Utility function to check sufficient n}
\usage{
check_histogram_n(accuracy, n, n_bins, epsilon, delta, alpha)
}
\description{
Utility function to check sufficient n
}

