% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_histogram_n}
\alias{check_histogram_n}
\title{Histogram N check}
\usage{
check_histogram_n(accuracy, n, n_bins, epsilon, delta, alpha)
}
\arguments{
\item{accuracy}{A numeric vector of length one representing the accuracy of 
the noisy estimate}

\item{n}{A numeric vector of length one specifying the number of
observations in in the data.}

\item{n_bins}{A numeric vector of length one specifying the number of cells 
in which to tabulate values.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{delta}{The probability of an arbitrary leakage of information from 
the data. Should be of length one and should be a very small value.}

\item{alpha}{A numeric vector of length one specifying the numeric 
statistical significance level.}
}
\value{
A logical vector indicating whether the number of observations is 
   sufficient to provide desired privacy and accuracy with the given
   parameters.
}
\description{
Utility function to check sufficient N in data.
}
