% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_variable_type}
\alias{check_variable_type}
\title{Utility function to check type of variable is within set of acceptable types}
\usage{
check_variable_type(type, in_types)
}
\arguments{
\item{type}{String specifying the type of the variable}

\item{in_types}{Vector of acceptable types 

Verifies that the variable is an element in the set of acceptable types}
}
\description{
Utility function to check type of variable is within set of acceptable types
}
\examples{
check_variable_type(type='Numeric', in_types=c('Numeric', 'Factor'))
}

