% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-validation.R
\name{checkDelta}
\alias{checkDelta}
\title{Check delta parameter}
\usage{
checkDelta(mechanism, delta)
}
\arguments{
\item{mechanism}{The mechanism chosen by determineMechanism}

\item{delta}{The delta value entered by the user, may be NULL}
}
\value{
The value that will be used as the delta value for the statistic, may be NULL.
}
\description{
This method is to send the user a warning message if they entered a delta value
that will not be used.
}
\details{
If the mechanism is NOT the stabiltiy mechanism and the user entered a delta value, 
send a warning message saying the delta value with not be used and set the delta 
value to NULL. If the stability mechanism is being used and the user entered a 
delta value, set it as the delta value (the value of delta will be checked in the 
stability mechanism). If the stability mechanism is being used and the user did 
not enter a delta value, set the delta to the default value (2^-30).
}
