% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-validation.R,
%   R/utilities-input-validation.R
\name{checkEpsilon}
\alias{checkEpsilon}
\title{Epsilon Parameter Check}
\usage{
checkEpsilon(epsilon, expectedLength = 1)

checkEpsilon(epsilon, expectedLength = 1)
}
\arguments{
\item{epsilon}{A vector, that ought to be positive and of length one.}

\item{expectedLength}{Integer value. Specifies how long the output ought to be.}

\item{epsilon}{A vector, that ought to be positive.}

\item{multipleEps}{A boolean value. If TRUE, multiple epsilon paramaters may be defined. Default to FALSE.}
}
\value{
The supplied epsilon if acceptable, otherwise an error 
   message interupts.

The supplied epsilon if acceptable, otherwise an error 
   message interupts.
}
\description{
Utility function for checking that epsilon is acceptably defined.

Utility function for checking that epsilon is acceptably defined.
}
\examples{

checkEpsilon(0.1)
\dontrun{checkEpsilon(-2)}
\dontrun{checkEpsilon(c(0.1,0.5))}
}
