% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-validation.R
\name{checkEpsilon}
\alias{checkEpsilon}
\title{Epsilon Parameter Check}
\usage{
checkEpsilon(epsilon)
}
\arguments{
\item{epsilon}{A vector, that ought to be positive and of length one.}
}
\value{
The supplied epsilon if acceptable, otherwise an error 
   message interupts.
}
\description{
Utility function for checking that epsilon is acceptably defined.
}
\examples{

checkEpsilon(0.1)
\dontrun{checkEpsilon(-2)}
\dontrun{checkEpsilon(c(0.1,0.5))}
}
