% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{checkImputationRange}
\alias{checkImputationRange}
\title{Error check imputation range for numeric or integer variables}
\usage{
checkImputationRange(imputationRange, rng, var.type)
}
\arguments{
\item{imputationRange}{The imputation range entered by the user}

\item{rng}{The data range entered by the user}

\item{var.type}{The variable type for the histogram data}
}
\value{
the imputation range that will be used for `fillMissing()`.
}
\description{
Check that the entered imputation range is within the entered data range. If yes, return
the entered imputation range, which will be used as the imputation range for the call
to the utility function `fillMissing()`. If not, return the data range. 
If the imputation range is NULL, default to the data range. The data range may be NULL,
in which case the stability mechanism will be used to determine the data range and
imputation range.
}
