% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkN}
\alias{checkN}
\title{Check validity of n}
\usage{
checkN(n, expectedLength = 1, emptyOkay = FALSE)
}
\arguments{
\item{n}{the input n (often number of data points in data set) from the user, or an array of n's from the user}

\item{expectedLength}{Positive integer. The expected length of n.}

\item{emptyOkay}{A boolean. True if NA or NULL values are allowed, FALSE otherwise.}
}
\value{
n, if n is a series of positive integers with expected length and only containing NAs or NULL values if allowed.
}
\description{
n(s) should always be a positive whole number, check the user's input
}
