% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkNumeric}
\alias{checkNumeric}
\title{Check if input is numeric type}
\usage{
checkNumeric(n)
}
\arguments{
\item{n}{Input value of arbitrary type.}
}
\value{
n or errors.
}
\description{
Helper function that generates error message if non-numeric typed value \code{n}
is passed as input, and otherwise returns \code{n}.
}
