% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkrange}
\alias{checkrange}
\title{Utility function for checking that range is ordered pair}
\usage{
checkrange(range)
}
\arguments{
\item{range}{A vector, that ought to be an ordered pair}
}
\value{
An ordered pair

Checks if a supplied range is an ordered pair.  Coerces any vector of length two
  or greater into an ordered pair, and issues an error for shorter vectors.
}
\description{
Utility function for checking that range is ordered pair
}
\examples{
checkrange(1:3)
\dontrun{checkrange(1)}
}

