% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkRange}
\alias{checkRange}
\title{Range Parameter Check}
\usage{
checkRange(rng)
}
\arguments{
\item{rng}{A numeric vector of length two, that ought to be an 
ordered pair.}
}
\value{
An ordered pair.
}
\description{
Checks if a supplied range is an ordered pair. Coerces any vector of length 
   two or greater into an ordered pair, and issues an error for
   shorter vectors.
}
\examples{

checkRange(c(1,3))
checkRange(1:3)
\dontrun{checkRange(1)}
}
