% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-input-validation.R
\name{checkRange1D}
\alias{checkRange1D}
\title{One-Dimensional Range Parameter Check}
\usage{
checkRange1D(rng, varType, emptyOkay = FALSE)
}
\arguments{
\item{rng}{Range. A numeric vector that ought to be an ordered pair.}

\item{varType}{String describing the variable type. One of 'logical' or 'numeric'.}

\item{emptyOkay}{Boolean. TRUE if a null or NA range is allowed, default to FALSE.}
}
\value{
Vector of length 2 or NULL
}
\description{
Helper function for checkRange. Checks if a supplied range is an ordered pair.
Coerces any vector of length two or greater into an ordered pair, and issues
an error for shorter vectors.
}
\details{
If varType is 'logical', range is coerced to c(0,1).

If emptyOkay, no error will be raised if the input range is null or NA, and an NA value will be returned.
}
