% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{covariance.postLinearRegression}
\alias{covariance.postLinearRegression}
\title{Function to perform linear regression using private release of covariance matrix}
\usage{
covariance.postLinearRegression(release, n, intercept, formulae)
}
\arguments{
\item{release}{Differentially private release of elements in lower triangle 
of covariance matrix.}

\item{n}{A numeric vector of length one specifying the number of
observations in the data frame.}

\item{intercept}{Logical indicating whether the intercept is included in 
\code{release}.}

\item{formulae}{A list of the regression equations to be performed on the 
covariance matrix.}
}
\value{
Linear regression coefficients and standard errors for all specified
   \code{formulae}.
}
\description{
Function to perform linear regression using private release of covariance matrix
}
