% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{covariance.sensitivity}
\alias{covariance.sensitivity}
\title{Function to get the sensitivity of the covariance matrix}
\usage{
covariance.sensitivity(n, rng, intercept)
}
\arguments{
\item{n}{A numeric vector of length one specifying the number of
observations in the data frame.}

\item{rng}{A numeric matrix of 2-tuples with the lower and upper bounds for
each of P variables in the data frame, dimensions Px2.}

\item{intercept}{A logical vector of length one indicating whether an 
intercept should be added prior to evaluating the inner product x'x.}
}
\value{
The sensitivity of the data frame for which the covariance matrix
  is being calculated.
}
\description{
Function to get the sensitivity of the covariance matrix
}
