% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-categorical.R
\name{createfields}
\alias{createfields}
\title{Create fields that are known based on type}
\usage{
createfields(v, r, varname)
}
\arguments{
\item{v}{an empty list needing initialization}

\item{r}{a release object for that variable}

\item{varname}{name of variable}
}
\description{
Create fields that are known based on type
}
