% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dlap}
\alias{dlap}
\title{Probability density for Laplace distribution}
\usage{
dlap(x, mu = 0, b = 1)
}
\arguments{
\item{x}{numeric, value}

\item{mu}{numeric, center of the distribution}

\item{b}{numeric, spread}
}
\value{
Density for elements of x
}
\description{
Probability density for Laplace distribution
}
\examples{

x <- seq(-3, 3, length.out=61)
dlap(x)
}
