% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-covariance.R
\name{dp.covariance}
\alias{dp.covariance}
\title{Function to evaluate the covariance matrix from input matrix and specify parameters for post-processing}
\usage{
dp.covariance(x, n, rng, epsilon, columns, intercept, formulae)
}
\arguments{
\item{x}{Numeric data frame}

\item{n}{Integer indicating the number of observations}

\item{rng}{Numeric matrix of 2-tuples with the lower and upper bounds for each of P variables, dimensions Px2}

\item{epsilon}{Numeric differential privacy parameter epsilon}

\item{columns}{Character vector indicating columns in \code{x}}

\item{intercept}{Logical indicating whether an intercept should be added prior to evaluating the inner product x'x}

\item{formulae}{something here}
}
\value{
List containing covariance matrix of the input data and postprocessing parameters
}
\description{
Function to evaluate the covariance matrix from input matrix and specify parameters for post-processing
}
