% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{dp.histogram}
\alias{dp.histogram}
\title{Function to evaluate a histogram and specify arguments}
\usage{
dp.histogram(x, var.type, stability, bins, n.bins, n, sensitivity, epsilon)
}
\arguments{
\item{x}{Vector of categorical or numeric values}

\item{var.type}{Character string indicating the variable type}

\item{stability}{Logical indicating if the stability mechanism is to be used}

\item{bins}{Vector of bins}

\item{n.bins}{Integer indicating the number of bins}

\item{n}{Integer indicating the number of observations in \code{x}}
}
\value{
List with the true value of the statistic and arguments to be passed to other functions
}
\description{
Function to evaluate a histogram and specify arguments
}
