% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{dp.histogram}
\alias{dp.histogram}
\title{DP histogram}
\usage{
dp.histogram(x, var.type, stability, bins, n.bins, n, sensitivity, epsilon)
}
\arguments{
\item{x}{A vector of categorical or numeric values.}

\item{var.type}{A character vector specifying variable type of \code{x}.}

\item{stability}{A logical indicating if the stability mechanism is to 
be used.}

\item{bins}{A vector of bins for which values are counted. Required for 
categorical types.}

\item{n.bins}{An integer specifying the number of cells in which to tabulate 
values in x. Ignored if \code{var.type \%in\% c('factor', 'categorical')}}

\item{n}{A numeric vector of length one specifying the number of
observations in \code{x}.}

\item{sensitivity}{The difference of the range of \code{x} divided 
by \code{n}.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}
}
\value{
A list with the true value of the statistic and arguments to be 
   passed to other functions.
}
\description{
Function to evaluate a true histogram and specify arguments for differentially private
histogram release.
}
