% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{dp.mean}
\alias{dp.mean}
\title{Function to evaluate the mean and specify parameters for mean functions}
\usage{
dp.mean(x, var.type, n, sensitivity, epsilon)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
List with fields `name` specifying the statistic and `stat` with the value of the statistic
}
\description{
Function to evaluate the mean and specify parameters for mean functions
}
