% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{dp.mean}
\alias{dp.mean}
\title{Function to evaluate the mean and specify parameters for mean functions}
\usage{
dp.mean(x, var.type, n, sensitivity, epsilon)
}
\arguments{
\item{x}{Numeric vector}

\item{var.type}{Character string indicating variable type}

\item{n}{The number of observations in vector x}

\item{sensitivity}{Numeric, the sensitivity of the estimate}

\item{epsilon}{Numeric, epsilon parameter for differential privacy}
}
\value{
List with fields `name` specifying the statistic and `stat` with the value of the statistic
}
\description{
Function to evaluate the mean and specify parameters for mean functions
}
