% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-variance.R
\name{dp.variance}
\alias{dp.variance}
\title{Function to evaluate true variance & specify post-processing parameters}
\usage{
dp.variance(x, var.type, n, sensitivity, epsilon)
}
\arguments{
\item{x}{Numeric vector}

\item{var.type}{Character string indicating variable type}

\item{n}{Integer indicating number of observations in \code{x}}

\item{sensitivity}{Numeric, the sensitivity of the estimate}

\item{epsilon}{Numeric, epsilon parameter for differential privacy}
}
\value{
List with fields providing values for the true statistic and post-processing
}
\description{
Function to evaluate true variance & specify post-processing parameters
}
