% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\docType{class}
\name{dpHistogram-class}
\alias{dpHistogram-class}
\alias{dpHistogram}
\title{Differentially private histogram}
\arguments{
\item{varType}{Character, the variable type.}

\item{variable}{Character, the variable name in the data frame.}

\item{n}{Integer, the number of observations.}

\item{epsilon}{Numeric, the privacy loss parameter.}

\item{accuracy}{Numeric, the desired accuracy of the query.}

\item{rng}{Numeric, a priori estimate of the lower and upper bounds of a
variable taking numeric values. Ignored for categorical types.}

\item{bins}{Character, the available bins or levels of a categorical variable.}

\item{nBins}{Integer, the number of bins to release.}

\item{granularity}{Numeric, the width of each histogram bin (i.e. the inverse of `nBins`). Used 
to calculate histogram bins in comination with `rng`.}

\item{alpha}{Numeric, level of statistical significance, default 0.05.}

\item{delta}{Numeric, probability of privacy loss beyond \code{epsilon}.}

\item{imputeRng}{Numeric, a 2-tuple indicating the lower and upper bounds of the range from which NA
values in numeric or integer-type variables should be imputed}

\item{imputeBins}{Character (or numeric for logical variables), a list of bins from which NA values
values in character-type variables should be imputed}

\item{impute}{Boolean, a boolean value indicating if logical-type variables should have NA values
imputed or not. If true, a logical variable histogram will have 2 bins, 0 and 1. If false, the
histogram will have 3 bins: 0, 1, and NA.}

\item{nBoot}{Numeric, the number of bootstrap iterations to do for bootstrapping (not used for version 1 release)}
}
\description{
Differentially private histogram
}

