% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\docType{class}
\name{dpHistogram-class}
\alias{dpHistogram-class}
\alias{dpHistogram}
\title{Differentially private histogram}
\arguments{
\item{mechanism}{Character, the mechanism used to perturb histogram bins.}

\item{var.type}{Character, the variable type.}

\item{variable}{Character, the variable name in the data frame.}

\item{n}{Integer, the number of observations.}

\item{epsilon}{Numeric, the privacy loss parameter.}

\item{accuracy}{Numeric, the desired accuracy of the query.}

\item{rng}{Numeric, a priori estimate of the lower and upper bounds of a
variable taking numeric values. Ignored for categorical types.}

\item{bins}{Character, the available bins or levels of a categorical variable.
Ignored for numeric types.}

\item{n.bins}{Integer, the number of bins to release.}

\item{alpha}{Numeric, level of statistical significance, default 0.05.}

\item{delta}{Numeric, probability of privacy loss beyond \code{epsilon}.}

\item{error}{Numeric, error.}
}
\description{
Differentially private histogram
}

