% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-glm.R
\name{dpLogit}
\alias{dpLogit}
\title{Differentially private objective function for Logistic regression}
\usage{
dpLogit()
}
\value{
List with the name and objective function
}
\description{
Differentially private objective function for Logistic regression
}

