% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dpNoise}
\alias{dpNoise}
\title{Draw cryptographically secure random variates}
\usage{
dpNoise(n, scale, dist, seed = NULL)
}
\arguments{
\item{n}{Integer giving number of variates needed}

\item{scale}{Numeric scale for the distribution}

\item{dist}{Character specifying the distribution from which to draw the noise}

\item{seed}{Integer indicating a seed for R's PNRG, defaults to \code{NULL}}
}
\description{
Draw cryptographically secure random variates
}
\examples{
laplace_noise <- dpNoise(n=1000, scale=1, dist='laplace')
gaussian_noise <- dpNoise(n=1000, scale=1, dist='gaussian')
laplace_noise_repeatable <- dpNoise(n=1, scale=1, dist='laplace', seed=96845)
}
