% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dpUnif}
\alias{dpUnif}
\title{Differentially Private Uniform Draw}
\usage{
dpUnif(n, seed = NULL)
}
\arguments{
\item{n}{An integer giving number of variates needed.}

\item{seed}{An integer indicating a seed for R's PNRG, defaults to \code{NULL}.}
}
\value{
Random numeric vector of length \code{n} containing values between
   zero and one.

Draws secure random variates from the uniform distribution through \code{openssl}.
If a seed is provided, the \code{runif} function is used to draw the random variates.
}
\description{
Draw cryptographically secure random variates from a uniform distribution.
}
\examples{

uniform_secure <- dpUnif(n=1000)
uniform_repeatable <- dpUnif(n=1, seed=75436)
}
\seealso{
\code{\link{dpNoise}}
}
