% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-variance.R
\docType{class}
\name{dpVariance-class}
\alias{dpVariance-class}
\alias{dpVariance}
\title{Differentially private variance}
\arguments{
\item{mechanism}{Character, the privacy mechanism.}

\item{var.type}{Character, the R variable type. One of \code{c('numeric',
'integer', 'logical')}.}

\item{Variable}{Character, variable name.}

\item{n}{Integer, number of observations}

\item{rng}{Numeric, a priori estimate of the range}

\item{epsilon}{Numeric, privacy cost parameter}

\item{accuracy}{Numeric, accuracy guarantee given \code{epsilon}}

\item{impute.rng}{Numeric, range within which missing values are imputed. If \code{NULL},
the range provided in \code{rng} is used.}

\item{alpha}{Numeric, the level of statistical significance. Default 0.05.}
}
\description{
Differentially private variance
}

