% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-variance.R
\docType{class}
\name{dpVariance-class}
\alias{dpVariance-class}
\alias{dpVariance}
\title{Differentially private variance}
\arguments{
<<<<<<< HEAD
\item{mechanism}{Character, the privacy mechanism. For \code{dpVariance}, one
of \code{c('mechanismLaplace', 'mechanismSnapping')}.}

\item{var.type}{Character, the R variable type. One of \code{c('numeric',
=======
\item{varType}{Character, the R variable type. One of \code{c('numeric',
>>>>>>> b81cbb2da5cb28a55f31b1a53f7ccf3c07aa0a4b
'integer', 'logical')}.}

\item{variable}{Character, variable name of the variable that you wish to find variance of.}

\item{n}{Integer, number of observations}

\item{rng}{Numeric, a priori estimate of the range}

\item{epsilon}{Numeric, privacy cost parameter}

\item{accuracy}{Numeric, accuracy guarantee given \code{epsilon}}

\item{imputeRng}{Numeric, range within which missing values are imputed. If \code{NULL},
the range provided in \code{rng} is used.}

\item{alpha}{Numeric, the level of statistical significance. Default 0.05.}

\item{gamma}{Numeric, used to provide high probability (1-gamma) guarantee that clamping bound does not bind. Default 0.05. Used only for \code{mechanismSnapping}.}
}
\description{
Differentially private variance
}

