% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{estBottomUp}
\alias{estBottomUp}
\title{Function to estimate a noisy binary tree from the terminal nodes}
\usage{
estBottomUp(tree, terminal.level.idx, n.nodes, sigma, inv.sigma.sq)
}
\arguments{
\item{tree}{Data frame with binary tree attributes and node values}

\item{terminal.level.idx}{Index of the first terminal leaf node}

\item{n.nodes}{Number of nodes in the binary tree}

\item{sigma}{Standard deviation of the noise used to perturb the estimates}

\item{inv.sigma.sq}{Inverse variance of the noise used in perturbing nodes}
}
\value{
Bottom-up estimate of noisy binary tree in vector form
}
\description{
Function to estimate a noisy binary tree from the terminal nodes
}

