% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{estEfficiently}
\alias{estEfficiently}
\title{Function to estimate a noisy binary tree efficiently using all available information in the tree}
\usage{
estEfficiently(tree, n, nNodes, sigma, invSigmaSq)
}
\arguments{
\item{tree}{Data frame with binary tree attributes and node values}

\item{n}{Number of observations in the vector represented by the binary tree}

\item{nNodes}{Number of nodes in the binary tree}

\item{sigma}{Standard deviation of the noise used to perturb the estimates}

\item{invSigmaSq}{Inverse variance of the noise used in perturbing nodes}
}
\value{
Efficient estimate of noisy binary tree in vector form
}
\description{
Function to estimate a noisy binary tree efficiently using all available information in the tree
}
