% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract.indices}
\alias{extract.indices}
\title{Function to obtain indices in data frame for dependent & independent variables from a formula}
\usage{
extract.indices(formula, data, intercept)
}
\arguments{
\item{formula}{Formula}

\item{data}{Data frame, in this case being the data frame of a private covariance matrix}

\item{intercept}{Logical indicating whether the intercept is included}
}
\value{
Named list with names corresponding to labels and locations (i.e., columns) for variables
 in the specification.
}
\description{
Function to obtain indices in data frame for dependent & independent variables from a formula
}
\examples{

y <- rnorm(100) * 2
x <- (y + rnorm(100)) > 0
data <- data.frame(cbind(y, x))
f <- as.formula('y ~ x')
extract.indices(f, data, FALSE)
}
