% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract.indices}
\alias{extract.indices}
\title{Extract regression indices}
\usage{
extract.indices(formula, data, intercept)
}
\arguments{
\item{formula}{An object of class 'formula' containing the desired 
regression formula.}

\item{data}{A numeric data frame with at least two columns.}

\item{intercept}{A logical vector indicating whether the intercept is 
included in \code{formula}.}
}
\value{
A named list with names corresponding to labels and locations 
   (i.e., columns) for variables in the specification.
}
\description{
Function to obtain indices in data frame for dependent & independent variables from a formula.
}
\examples{

y <- rnorm(100) * 2
x <- (y + rnorm(100)) > 0
data <- data.frame(cbind(y, x))
f <- as.formula('y ~ x')
extract.indices(f, data, FALSE)
}

