% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data-validation.R
\name{fillMissing}
\alias{fillMissing}
\title{Fill missing values}
\usage{
fillMissing(x, varType, imputeRng = NULL, categories = NULL)
}
\arguments{
\item{x}{Vector with missing values to impute}

\item{varType}{Character specifying the variable type}

\item{categories}{Set of possible categories from which to choose,
default NULL}

\item{lower}{Numeric lower bound, default NULL}

\item{upper}{Numeric upper bound, default NULL}
}
\value{
Vector \code{x} with missing values imputed
}
\description{
Impute uniformly in the range of the provided variable
}
\examples{

# numeric example
y <- rnorm(100)
y[sample(1:100, size=10)] <- NA
y_imputed <- fillMissing(x=y, varType='numeric', lower=-1, upper=1)

# categorical example
cats <- as.factor(c('dog', 'zebra', 'bird', 'hippo'))
s <- sample(cats, size=100, replace=TRUE)
s[sample(1:100, size=10)] <- NA
s_imputed <- fillMissing(x=s, varType='factor', categories=cats)

}
\seealso{
\code{\link{dpUnif}}
}
