% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fillMissing2d}
\alias{fillMissing2d}
\title{Fill missing values column-wise for matrix}
\usage{
fillMissing2d(x, var.type, impute.rng = NULL)
}
\arguments{
\item{x}{Numeric matrix with missing values}

\item{var.type}{Character specifying the variable type}

\item{impute.rng}{Px2 matrix where the pth row contains the range
within which the pth variable in x is imputed.}
}
\value{
Matrix \code{x} with missing values imputed
}
\description{
Impute uniformly in the range of the provided variable
}
\examples{

# numeric example
N <- 100
x1 <- x2 <- rnorm(N)
x1[sample(1:N, size=10)] <- NA
x2[sample(1:N, size=10)] <- NA
imp.rng <- matrix(c(-3, 3, -2, 2), ncol=2, byrow=TRUE)
df <- data.frame(x1, x2)
df_imputed <- fillMissing2d(x=df, var.type='numeric', impute.rng=imp.rng)

}
\seealso{
\code{\link{fillMissing}}
}
