% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getFuncArgs}
\alias{getFuncArgs}
\title{Utility function to match arguments of a function with list output of another function}
\usage{
getFuncArgs(output, target.func)
}
\arguments{
\item{output}{List with output of a function}

\item{target.func}{Character name of the function with arguments that need to be filled by output}
}
\value{
List of arguments and values needed for specification of \code{target.func}
}
\description{
Utility function to match arguments of a function with list output of another function
}

