% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{heavyhitters.getAccuracy}
\alias{heavyhitters.getAccuracy}
\title{Get the accuracy of heavyhitters statistic for a given value of epsilon}
\usage{
heavyhitters.getAccuracy(gap, epsilon, delta)
}
\arguments{
\item{gap}{Gap}

\item{epsilon}{Epsilon value for differential privacy}

\item{delta}{Delta value for differential privacy}
}
\value{
The accuracy guaranteed by the given epsilon
}
\description{
Get the accuracy of heavyhitters statistic for a given value of epsilon
}
\author{
Victor Balcer

The accuracy is interpreted as follows: The alpha value returned means that with
  probability 1 - beta, simultaneously for all t with min <= t <= max, the algorithm's
  estimate of the count in [min, t] is within alpha of the true value.
}
