% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{heavyhitters.getParameters}
\alias{heavyhitters.getParameters}
\title{Heavyhitters epsilon}
\usage{
heavyhitters.getParameters(gap, delta, alpha = 0.05)
}
\arguments{
\item{gap}{}

\item{delta}{A numeric vector representing the probability of an arbitrary
leakage of information from the data. Should be of length one 
and should be a very small value.}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
The epsilon value necessary to gaurantee the given accuracy.
}
\description{
Get the epsilon value necessary to guarantee a desired level of accuracy of a heavyhitters release
}
\author{
Victor Balcer
}
