% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{heavyhitters.getParameters}
\alias{heavyhitters.getParameters}
\title{Get the epsilon value necessary to guarantee a desired level of accuracy of a heavyhitters release}
\usage{
heavyhitters.getParameters(delta, beta = 0.05, gap)
}
\arguments{
\item{delta}{Delta value for DP}

\item{beta}{the true value is within the accuracy range (alpha)}
}
\value{
The epsilon value necessary to gaurantee the given accuracy
}
\description{
Get the epsilon value necessary to guarantee a desired level of accuracy of a heavyhitters release
}
\author{
Victor Balcer
}

