% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{heavyhitters.postNoteFailure}
\alias{heavyhitters.postNoteFailure}
\title{Heavyhitters failure}
\usage{
heavyhitters.postNoteFailure(failed)
}
\arguments{
\item{failed}{A logical vector.}
}
\value{
\code{failed}. If \code{TRUE}, the mechanism failed. If \code{FALSE},
   the function succeeded.
}
\description{
Function to see if the heavyhitters mechanism failed.
}
