% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{heavyhitters.release}
\alias{heavyhitters.release}
\title{Release differentially private heavyhitters}
\usage{
heavyhitters.release(x, var.type, epsilon, n, k, bins, delta = 1e-06)
}
\arguments{
\item{x}{A vector of categorical data (character, factor).}

\item{var.type}{A character vector specifying variable type of \code{x}.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{n}{A numeric vector of length one specifying the number of
observations in \code{x}.}

\item{k}{An integer querying the most common \code{k} categories.}

\item{bins}{A vector of categories from which the top \code{k} categories 
are evaluated.}

\item{delta}{A numeric vector representing the probability of an arbitrary
leakage of information from \code{x}. Should be of length one 
and should be a very small value. Default to 10^-6.}
}
\value{
A vector whose values are \code{k} most common categories, or a 
   failure message.
}
\description{
A function for privately releasing the top-k heavy hitters, that is,
 the k most common unique values.  For example, k=1 releases the mode.
}
\examples{

N <- 1000
epsilon <- 0.5
delta <- 1e-7
observed.levels <- bins <- c('a', 'b', 'c', 'd', 'e', 'f', 'g', 'h')
probs <- c(0.40, 0.25, 0.15, 0.10, 0.04, 0.03, 0.02, 0.01)
y <- sample(observed.levels, size=N, prob=probs, replace=TRUE)
release <- heavyhitters.release(y, 'character', epsilon, delta, N, k=3, bins=bins)
}
