% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-heavyhitters.R
\name{heavyhitters.release}
\alias{heavyhitters.release}
\title{Release differentially private most common values.}
\usage{
heavyhitters.release(x, epsilon, delta, k = 1, symb = NA)
}
\arguments{
\item{x}{A vector of the data}

\item{epsilon}{Epsilon value for differential privacy}

\item{delta}{Delta value for differential privacy}

\item{k}{top-k elements to return}

\item{symb}{Symbol to return on failure (default : NA)}
}
\value{
A vector whose values are k most common values or a failure return symbol.
}
\description{
Release differentially private most common values.
}
\examples{
n <- 10000
range <- c(0,20)
x <- rbinom(n, size=max(range), prob=0.7)
heavyhitters.release(x=x, epsilon=.1, delta=.0000001)
}
\author{
Victor Balcer

Contains functions for privately releasing the top-k heavy hitters, that is,
 the k most common unique values.  For example, k=1 releases the mode.
}

