% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.getAccuracy}
\alias{histogram.getAccuracy}
\title{Accuracy of release}
\usage{
histogram.getAccuracy(mechanism, n_bins, n, epsilon, delta = 2^-30,
  beta = 0.05, error = 1e-09)
}
\arguments{
\item{mechanism}{Specify the histogram mechanism}

\item{n_bins}{Number of cells in which to tabulate values}

\item{epsilon}{Epsilon value for differential privacy}

\item{delta}{Delta value for differential privacy}
}
\description{
Accuracy of release
}

