% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{histogram.getAccuracy}
\alias{histogram.getAccuracy}
\title{Histogram accuracy}
\usage{
histogram.getAccuracy(mechanism, n.bins, n, epsilon, delta = 10^-6,
  alpha = 0.05)
}
\arguments{
\item{mechanism}{A string indicating the mechanism that will be used to construct the histogram}

\item{n.bins}{A numeric vector of length one specifying the number of cells 
in which to tabulate values.}

\item{n}{A numeric vector of length one specifying the number of
observations in the data.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{delta}{The probability of an arbitrary leakage of information from 
the data. Should be of length one and should be a very small value. 
Default to 10^-6.}

\item{alpha}{A numeric vector of length one specifying the numeric 
statistical significance level. Default to 0.05.}
}
\value{
Accuracy guarantee for histogram release, given epsilon.
}
\description{
Determine accuracy of histogram release, given epsilon and delta, for the differentially 
private histogram release.
}
\details{
In differential privacy, "accuracy" is defined as the threshold value above which a given value 
is "significantly different" from the expected value. Mathematically, this is written as:
\deqn{\alpha = Pr[Y > a]}
Where \eqn{\alpha} is the statistical significance level, \eqn{a} is the accuracy, and \eqn{Y} is a random 
variable following the Laplace distribution.

In other words, the accuracy is a threshold value, above which a value is "unexpected". We call
this "accuracy" because most values will be below this threshold, and thus are "accurate" enough
to be considered "expected" values.

Deriving the accuracy formula:

\enumerate{
    \item The probability density function (PDF) \eqn{f(x)} of the Laplace distribution is:\cr
          \eqn{f(x) = {1 / 2\lambda} * e^{-|x-\mu| / \lambda}}
    \item For differential privacy calculations \eqn{\mu = 0}, because the "average" amount of noise added is \eqn{0} noise
    \item For differential privacy, "accuracy" is defined as the magnitude of difference between the expected and observed values, so the distrubution is "folded", or doubled, because we are only concerned with the absolute value
    \item So we can consider the differentially private PDF \eqn{g(Y)} to be:\cr
    \eqn{g(y) = {1 / \lambda} * e^{-y / \lambda}}
    \item Using \eqn{\alpha = Pr[Y > a]} and the PDF, we can solve for \eqn{a} and plug in \eqn{\lambda = 2 / \epsilon}
    \item We end up with the accuracy formula: \eqn{a = {2 / \epsilon} * ln(1 / \alpha)}
    \item The accuracy formula for the stability mechanism is dervied by adding the accuracy formula above to the accuracy threshold (which is the worst-case potentially added noise in the stability mechanism): \eqn{{2 / \epsilon} * ln(2 / \delta)+1}
}
}
\references{
S. Vadhan The Complexity of Differential Privacy, Section 3.3 Releasing Stable Values p.23-24. March 2017.
}
