% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.getAccuracy}
\alias{histogram.getAccuracy}
\title{Accuracy of release}
\usage{
histogram.getAccuracy(n.bins, n, epsilon, stability, delta = 2^-30,
  alpha = 0.05, error = 0.000000001)
}
\arguments{
\item{n.bins}{Integer indicating number of cells in which to tabulate values}

\item{n}{Integer indicating number of observations}

\item{epsilon}{Numeric epsilon value for differential privacy}

\item{stability}{Logical indicating whether stability mechanism is used}

\item{delta}{Numeric delta value for differential privacy, fixed}

\item{alpha}{Numeric statistical significance level, fixed}

\item{error}{Numeric, fixed}
}
\value{
Accuracy
}
\description{
Accuracy of release
}
