% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.getAccuracy}
\alias{histogram.getAccuracy}
\title{Histogram accuracy}
\usage{
histogram.getAccuracy(n.bins, n, epsilon, stability, delta = 2^-30,
  alpha = 0.05, error = 1e-09)
}
\arguments{
\item{n.bins}{A numeric vector of length one specifying the number of cells 
in which to tabulate values.}

\item{n}{A numeric vector of length one specifying the number of
observations in in the data.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{stability}{A logical vector indicating whether the stability 
mechanism is used.}

\item{delta}{The probability of an arbitrary leakage of information from 
the data. Should be of length one and should be a very small value. 
Default to 10^-6.}

\item{alpha}{A numeric vector of length one specifying the numeric 
statistical significance level. Default to 0.05.}

\item{error}{The error term of the statistical significance level. Default
to 1e-9.}
}
\value{
Accuracy guarantee for histogram release, given epsilon.
}
\description{
Determine accuracy of histogram release, given epsilon and delta, for the differentially 
private histogram release.
}
