% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.getCI}
\alias{histogram.getCI}
\title{Histogram confidence interval}
\usage{
histogram.getCI(release, n.bins, n, accuracy)
}
\arguments{
\item{release}{A numeric vector with a noisy estimate of bin counts.}

\item{n.bins}{A numeric vector of length one specifying the number of cells 
in which to tabulate values.}

\item{n}{A numeric vector of length one specifying the number of
observations in in the data.}

\item{accuracy}{A numeric vector representing the accuracy needed to 
guarantee (percent).}
}
\value{
Confidence interval for the noisy counts in each bin.
}
\description{
Return the confidence interval for the differentially private histogram release given the
accuracy.
}
