% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.getJSON}
\alias{histogram.getJSON}
\title{JSON doc for histogram}
\usage{
histogram.getJSON(output.json = TRUE)
}
\arguments{
\item{output.json}{something here}
}
\value{
JSON for histogram function
}
\description{
JSON doc for histogram
}
