% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-histogram.R
\name{histogram.getParameters}
\alias{histogram.getParameters}
\title{Histogram Epsilon}
\usage{
histogram.getParameters(n.bins, n, accuracy, stability, delta = 2^-30,
  alpha = 0.05, error = 1e-09)
}
\arguments{
\item{n.bins}{A numeric vector of length one specifying the number of cells 
in which to tabulate values.}

\item{n}{A numeric vector of length one specifying the number of
observations in in the data.}

\item{accuracy}{A numeric vector representing the accuracy needed to 
guarantee (percent).}

\item{delta}{The probability of an arbitrary leakage of information from 
the data. Should be of length one and should be a very small value. 
Default to 10^-6.}

\item{alpha}{A numeric vector of length one specifying the numeric 
statistical significance level. Default to 0.05.}

\item{error}{}
}
\value{
Differential privacy parameter epsilon
}
\description{
Function to find the epsilon value necessary to meet a desired level of accuracy.
}
