% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-histogram.R
\name{histogram.postHerfindahl}
\alias{histogram.postHerfindahl}
\title{Histogram Herfindahl Index}
\usage{
histogram.postHerfindahl(release, n)
}
\arguments{
\item{release}{A numeric vector with a noisy estimate of bin counts.}

\item{n}{A numeric vector of length one specifying the number of
observations in in the data.}
}
\value{
Herfindahl index.
}
\description{
Produce differentially private Herfindahl index for categorical types of data.
}
