% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism-laplace.R
\name{Laplace Mechanism}
\alias{Laplace Mechanism}
\title{Laplace Mechanism}
\arguments{
\item{fun}{function of input x to add Laplace noise to.}

\item{x}{input that function fun will be evaluated on.}

\item{sens}{sensitivity of fun. Sensitivity is defined in above citation.}

\item{postFun}{post-processing function. Takes differentially private release as input
and returns some form of output in principal based on the differentially private release.}

\item{...}{any additional (optional) parameters}
}
\value{
result of post-processing on input function "fun" evaluated on database "x", assuming sensitivity of fun is "sens".
}
\description{
Differentially private evaluation of input function "fun" with sensitivity "sens" on input data 
"x" using the Laplace mechanism.
}
\examples{
laplace_histogram <- mechanismLaplace$evaluate(fun.hist, data[, variable], 2, dpHistogram$postProcess)
laplace_mean <- mechanismLaplace$evaluate(mean, data[, variable], sens, dpMean$postProcess)

}
\references{
C. Dwork, A. Roth The Algorithmic Foundations of Differential Privacy, Chapter 3.3 The Laplace Mechanism p.30-37. August 2014.
}
