% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.getAccuracy}
\alias{mean.getAccuracy}
\title{Describe Here}
\usage{
\method{mean}{getAccuracy}(epsilon, n, alpha = 0.05)
}
\arguments{
\item{epsilon}{Privacy parameter epsilon}

\item{n}{Number of observations}

\item{alpha}{The statistical significance level}
}
\value{
Accuracy guarantee for mean release given epsilon
}
\description{
Describe Here
}
