% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mean.R
\name{mean.getCI}
\alias{mean.getCI}
\title{Mean confidence interval}
\usage{
\method{mean}{getCI}(release, epsilon, sensitivity, alpha = 0.05)
}
\arguments{
\item{release}{Differentially private release of a mean.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{sensitivity}{The difference of the range of the data divided 
by \code{n}.}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
Confidence bounds for differentially private mean release.
}
\description{
Return the confidence interval for the differentially private mean release given the
accuracy.
}
