% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.getParameters}
\alias{mean.getParameters}
\title{Describe Here}
\usage{
\method{mean}{getParameters}(accuracy, n, alpha = 0.05)
}
\arguments{
\item{accuracy}{The accuracy we need to guarantee (percent)}

\item{n}{The number of samples}

\item{alpha}{The statistical signifcance level}
}
\value{
The scalar epsilon necessary to guarantee the accuracy needed
}
\description{
Describe Here
}
