% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.getParameters}
\alias{mean.getParameters}
\title{Mean epsilon}
\usage{
\method{mean}{getParameters}(accuracy, n, alpha = 0.05, rng)
}
\arguments{
\item{accuracy}{A numeric vector representing the accuracy needed to 
guarantee (percent).}

\item{n}{A numeric vector of length one specifying the number of
observations in the vector calculating the mean for.}

\item{alpha}{A numeric vector specifying the statistical significance level.}
}
\value{
The scalar epsilon necessary to guarantee the needed accuracy.
}
\description{
Function to find the epsilon value necessary to meet a desired level of 
accuracy for a mean release.
}
