% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic-mean.R
\name{mean.postMedian}
\alias{mean.postMedian}
\title{Postprocessed median for logical variables}
\usage{
\method{mean}{postMedian}(release)
}
\arguments{
\item{release}{Differentially private release of a mean for a logical
variable.}
}
\value{
Median of the logical variable
}
\description{
Calculates the median of the differentially private mean from a
logical variable.
}
