% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.postStandardDeviation}
\alias{mean.postStandardDeviation}
\title{Postprocessed Standard Deviation for Logical Variables}
\usage{
\method{mean}{postStandardDeviation}(release)
}
\arguments{
\item{release}{Differentially private release of a mean for a logical 
variable.}
}
\value{
Standard deviation of \code{release}.
}
\description{
Postprocessed Standard Deviation for Logical Variables
}
