% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-mean.R
\name{mean.release}
\alias{mean.release}
\title{Function for differentially private release of mean}
\usage{
\method{mean}{release}(x, var.type, n, epsilon, rng)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{The number of samples}

\item{epsilon}{Numeric}

\item{range}{An a priori estimate of the range}
}
\value{
Differentially private release of mean of vector x
}
\description{
Function for differentially private release of mean
}
\examples{

n <- 1000
x_num <- runif(n)
x_int <- as.integer(round(x_num * 100))
x_bool <- x_num >= 0.5
x_dich <- ifelse(x_bool, 3.483, -9.657)

r_num <- mean.release(x=x_num, var.type='numeric', epsilon=0.5, n=n, range=c(0, 1))
r_int <- mean.release(x=x_int, var.type='integer', epsilon=0.5, n=n, range=c(5, 95))
r_bool <- mean.release(x=x_bool, var.type='logical', epsilon=0.5, n=n, range=c(0, 1))
r_dich <- mean.release(x=x_dich, var.type='logical', epsilon=0.5, n=n, range=c(-9.657, 3.483))
}
