% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{mechanism.gaussian}
\alias{mechanism.gaussian}
\title{Gaussian mechanism}
\usage{
mechanism.gaussian(fun, x, var.type, rng, sensitivity, epsilon, delta,
  postlist = NULL, ...)
}
\arguments{
\item{fun}{A user supplied function, or string naming a function.  Must accept only one argument, named \code{x}}

\item{x}{A vector of data to run the function on}

\item{var.type}{Character string indicating variable of vector \code{x}}

\item{rng}{An a priori estimate of the range}

\item{sensitivity}{Numeric, the sensitivity of the estimate}

\item{epsilon}{Numeric, epsilon parameter for differential privacy}

\item{delta}{something here}

\item{postlist}{List with name, function pairs for post-processing statistics}

\item{...}{Other arguments passed to \code{fun}}
}
\value{
something here
}
\description{
Gaussian mechanism
}
