% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{mechanism.laplace}
\alias{mechanism.laplace}
\title{LaPlace mechanism for releasing differentially private queries}
\usage{
mechanism.laplace(fun, x, var_type, range, sensitivity, epsilon,
  levels = NULL, ...)
}
\arguments{
\item{fun}{A user supplied function, or string naming a function.  Must accept only one argument, named \code{x}.}

\item{x}{A vector of data to run the function on.}

\item{var_type}{Data type of vector x}

\item{range}{An a priori estimate of the range}

\item{sensitivity}{numeric}

\item{epsilon}{numeric}

\item{n}{The number of samples}
}
\value{
Differentially private release of function \code{fun} on data \code{x}
}
\description{
LaPlace mechanism for releasing differentially private queries
}
\examples{
n <- 1000
range <- c(0,1)
x <- runif(n, min=min(range), max=max(range))
sensitivity <- diff(range)/n
mechanism.laplace(mean,x=x,sensitivity=sensitivity, epsilon=.1)
}

