% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{mechanism.laplace}
\alias{mechanism.laplace}
\title{LaPlace mechanism for releasing differentially private queries}
\usage{
mechanism.laplace(fun, x, var.type, rng, sensitivity, epsilon,
  postlist = NULL, ...)
}
\arguments{
\item{fun}{A user supplied function, or string naming a function.  Must accept only one argument, named \code{x}}

\item{x}{A vector of data to run the function on}

\item{var.type}{Character string indicating variable of vector \code{x}}

\item{rng}{An a priori estimate of the range}

\item{sensitivity}{Numeric, the sensitivity of the estimate}

\item{epsilon}{Numeric, epsilon parameter for differential privacy}

\item{postlist}{List with name, function pairs for post-processing statistics}

\item{...}{Other arguments passed to \code{fun}}
}
\value{
Differentially private release of function \code{fun} on data \code{x}
}
\description{
LaPlace mechanism for releasing differentially private queries
}
\examples{
n <- 1000
rng <- c(0,1)
x <- runif(n, min=min(rng), max=max(rng))
sensitivity <- diff(rng) / n
mechanism.laplace(fun=dp.mean, x=x, var.type='numeric', rng=rng, sensitivity=sensitivity, epsilon=0.5, n=n)
}
