% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-covariance.R
\name{mpinv}
\alias{mpinv}
\title{Moore Penrose Inverse Function}
\usage{
mpinv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{A numeric, symmetric covariance matrix.}

\item{tol}{Convergence requirement.}
}
\description{
Moore Penrose Inverse Function
}
\references{
Gill, Jeff, and Gary King. "What to do when your Hessian is not invertible: Alternatives to model respecification in nonlinear estimation." Sociological methods & research 33, no. 1 (2004): 54-87.

Generate the Moore-Penrose pseudoinverse matrix of \code{X}.
}
