% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{postprocess}
\alias{postprocess}
\title{Cycle through available postprocessing functions for a released statistic}
\usage{
postprocess(out, postlist, ...)
}
\arguments{
\item{out}{A list containing differentially private released statistic, and 
mechanism and statistic names.}

\item{postlist}{A list with names, function pairs for post-processing statistics.}

\item{...}{Other arguments.}
}
\value{
Postprocessed statistics.
}
\description{
Cycle through available postprocessing functions for a released statistic
}
