\name{PSIlence-package}
\alias{PSIlence-package}
\alias{PSIlence}
\docType{package}
\title{
\packageTitle{PSIlence}
}
\description{
\packageDescription{PSIlence}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PSIlence}
\packageIndices{PSIlence}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PSIlence}

Maintainer: \packageMaintainer{PSIlence}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
