% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-quantiles.R
\name{quantile.getAccuracy}
\alias{quantile.getAccuracy}
\title{Get the accuracy of quantile statistic for a given value of epsilon}
\usage{
\method{quantile}{getAccuracy}(epsilon, beta, range, gran, n)
}
\arguments{
\item{epsilon}{Epsilon value for DP}

\item{beta}{The true value is within the accuracy range with
probability 1-beta}

\item{range}{The range of the universe as a vector (min, max)}

\item{gran}{The granularity of the universe between the min and max}

\item{n}{Number of observations of the variable}
}
\value{
The accuracy guaranteed by the given epsilon
}
\description{
Get the accuracy of quantile statistic for a given value of epsilon
}
\author{
Nathan Manohar

The accuracy is interpreted as follows: The alpha value returned means that with 
  probability 1 - beta, simultaneously for all t with min <= t <= max, the algorithm's 
  estimate of the count in [min, t] is within alpha of the true value.
}

