% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-quantiles.R
\name{quantile.getAccuracy}
\alias{quantile.getAccuracy}
\title{Get the accuracy of quantile statistic for a given value of epsilon}
\usage{
\method{quantile}{getAccuracy}(epsilon, n, universe.size, alpha = 0.05)
}
\arguments{
\item{epsilon}{Numeric, epsilon value for DP}

\item{n}{Integer, number of observations}

\item{universe.size}{Integer, the universe size}

\item{alpha}{Numeric, statistical significance level}
}
\value{
The accuracy guaranteed by the given epsilon
}
\description{
Get the accuracy of quantile statistic for a given value of epsilon
}
\author{
Nathan Manohar

The accuracy is interpreted as follows: The alpha value returned means that with 
  probability 1 - beta, simultaneously for all t with min <= t <= max, the algorithm's 
  estimate of the count in [min, t] is within alpha of the true value.
}

