% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-quantiles.R
\name{quantile.getParameters}
\alias{quantile.getParameters}
\title{Quantiles epsilon}
\usage{
\method{quantile}{getParameters}(accuracy, n, universe.size, alpha = 0.05)
}
\arguments{
\item{accuracy}{A numeric vector representing the accuracy needed to 
guarantee (percent).}

\item{n}{A numeric vector of length one specifying the number of
observations in the data.}

\item{universe.size}{Integer, the universe size}

\item{alpha}{A numeric vector specifying the statistical significance level.
Default to 0.05.}
}
\value{
The epsilon value necessary to gaurantee the given accuracy.
}
\description{
Get the epsilon value necessary to guarantee a desired level of accuracy of a quantile release.
}
\author{
Nathan Manohar
}
