% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-quantiles.R
\name{quantile.getParameters}
\alias{quantile.getParameters}
\title{Get the epsilon value necessary to guarantee a desired level of accuracy of a quantile release}
\usage{
\method{quantile}{getParameters}(accuracy, n, universe.size, alpha = 0.05)
}
\arguments{
\item{accuracy}{Numeric, the accuracy parameter}

\item{n}{Integer, number of observations}

\item{universe.size}{Integer, the universe size}

\item{alpha}{Numeric, statistical significance level}
}
\value{
The epsilon value necessary to gaurantee the given accuracy
}
\description{
Get the epsilon value necessary to guarantee a desired level of accuracy of a quantile release
}
\author{
Nathan Manohar
}
