% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-quantiles.R
\name{quantile.release}
\alias{quantile.release}
\title{Release differentially private quantiles}
\usage{
\method{quantile}{release}(x, var.type, n, epsilon, rng, gran, cdf.step, ...)
}
\arguments{
\item{x}{A numeric or integer vector.}

\item{var.type}{A character vector specifying variable type of \code{x}.
Should be of length one.}

\item{n}{A numeric vector of length one specifying the number of
observations in \code{x}.}

\item{epsilon}{A numeric vector representing the epsilon privacy parameter.
Should be of length one and should be between zero and one.}

\item{rng}{A numeric vector specifying an a priori estimate of the range
of \code{x}. Should be of length two.}

\item{...}{additional arguments passed to \code{quantile.release}.}
}
\description{
Function to calculate a differentially private release of quantiles.
}
\examples{
 
}
