% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp-regression.R
\name{regression.release}
\alias{regression.release}
\title{Function to perform regression using the differentially private covariance matrix via the sweep operator}
\usage{
regression.release(formula, release, n, method = "linear")
}
\arguments{
\item{formula}{Formula, regression formula used on data}

\item{release}{Numeric, private release of covariance matrix}

\item{n}{Integer, indicating number of observations}

\item{method}{Character vector, the regression technique used}
}
\description{
Function to perform regression using the differentially private covariance matrix via the sweep operator
}
