% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rlap}
\alias{rlap}
\title{Random draw from Laplace distribution}
\usage{
rlap(mu = 0, b = 1, size = 1)
}
\arguments{
\item{mu}{numeric, center of the distribution}

\item{b}{numeric, spread}

\item{size}{number of observations}
}
\value{
Random draws from Laplace distribution
}
\description{
Random draw from Laplace distribution
}
\examples{
rlap(size=1000)
}
