% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-noise-generation.R
\name{sgn}
\alias{sgn}
\title{Sign function}
\usage{
sgn(x)
}
\arguments{
\item{x}{numeric, value or vector or values}
}
\value{
The sign of passed values
}
\description{
Function to determine what the sign of the passed values should be.
}
\examples{
sgn(rnorm(10))
}
