% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-noise-generation.R
\name{snappingNoise}
\alias{snappingNoise}
\title{Noise Generator for Snapping Mechanism}
\usage{
snappingNoise(true_val, n, sens, epsilon, min_B)
}
\arguments{
\item{true_val}{True value of the parameter.}

\item{n}{An integer giving number of variates needed.}

\item{sens}{Sensitivity of the function for which we are releasing a private estimate.}

\item{epsilon}{Desired level of differential privacy}

\item{min_B}{Minimum bound B such that the true value is guaranteed to be within [-B, B]

noise <- snappingNoise(true_val = 50, sens = 5e-4, epsilon = 1e-3, B = 200)}
}
\description{
Generate noise for use in Snapping Mechanism.
}
