% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanism-stability.R
\name{Stability Mechanism}
\alias{Stability Mechanism}
\title{Stability Mechanism}
\arguments{
\item{fun}{function of input x to add Laplace noise to.}

\item{x}{input that function fun will be evaluated on.}

\item{sens}{sensitivity of fun. Sensitivity is defined in above citation.}

\item{postFun}{post-processing function. Takes differentially private release as input
and returns some form of output in principal based on the differentially private release.}

\item{...}{any additional (optional) parameters}
}
\value{
result of post-processing on input function "fun" evaluated on database "x", assuming sensitivity of fun is "sens".
}
\description{
Differentially private evaluation of input function "fun" with sensitivity "sens" on input data 
"x" using the Stability mechanism. The Stability mechanism gives a stronger guarantee that the 
output of a query $f$ on a database $x$ will be close to the output of $f$ on any neighboring database
\eqn{\hat{x}}. For a histogram generated by the Stability mechanism, empty buckets will be removed and any
buckets with a count below an accuracy threshold will also be removed. Removing these buckets based on
low counts is what creates the added guarantee of privacy, by hiding data that could be different in
neighboring databases.
}
\details{
The accuracy threshold for removing bins with low counts is calculated by: \eqn{1+\frac{2*log(2/delta)}{epsilon}}.
}
\examples{
stability_histogram <- mechanismStability$evaluate(fun.hist, data[, variable], 2, dpHistogram$postProcess)
stability_mean <- mechanismStability$evaluate(mean, data[, variable], sens, dpMean$postProcess)

}
\references{
S. Vadhan The Complexity of Differential Privacy, Section 3.3 Releasing Stable Values p.23-24. March 2017.
}
