% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{treePostCDF}
\alias{treePostCDF}
\title{Function to truncate negative noisy node counts at zero}
\usage{
treePostCDF(release, rng, terminalIndex)
}
\arguments{
\item{release}{The differentially private noisy binary tree}

\item{rng}{An a priori estimate of the range of the vector
being represented as a binary tree}

\item{terminalIndex}{Vector of indices corresponding to the terminal
leaf nodes of the binary tree}

\item{release}{Efficient differentially private binary tree}
}
\value{
Noisy binary tree truncated at zero
Function to derive CDF from efficient terminal node counts

Differentially private estimate of the empirical cumulative
     distribution function
FIX THIS
}
\description{
Function to truncate negative noisy node counts at zero
}
