% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP-variance.R
\name{variance.release}
\alias{variance.release}
\title{Function for differentially private release of variance}
\usage{
variance.release(x, var.type, n, epsilon, rng)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Integer indicating number of observations in \code{x}}

\item{epsilon}{Numeric, epsilon parameter for differential privacy}

\item{rng}{Numeric 2-tuple indicating range of numeric variable}

\item{var.trpe}{Character string indicating variable type}
}
\value{
List output containing noisy estimate and post-processing values from the Laplace mechanism
}
\description{
Function for differentially private release of variance
}
