% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wBelow}
\alias{wBelow}
\title{Function to evaluate weights from the noise variance and standard errors in child nodes for the 
 node of a differentially private binary tree}
\usage{
wBelow(inv.sigma.sq, tree, idx)
}
\arguments{
\item{inv.sigma.sq}{Inverse variance of the noise used in perturbing nodes}

\item{tree}{Data frame with binary tree attributes and node values}

\item{idx}{Index of the node for which the weight is evaluated}
}
\value{
Weight
}
\description{
Function to evaluate weights from the noise variance and standard errors in child nodes for the 
 node of a differentially private binary tree
}
