% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tree.R
\name{wBelow}
\alias{wBelow}
\title{Recursive weight estimate from below}
\usage{
wBelow(tree)
}
\arguments{
\item{tree}{Tree, formatted as a list of arrays where the contents of the ith array in the list is the ith level of the tree.}
}
\value{
Single weight for each level of the tree.
}
\description{
See extra_docs/tree-post-processing for formula. This assumes that the variance is the same for every node at in the tree.
}
\examples{
t <- list(c(10), c(6,4), c(3,3,1,3))
w <- wBelow(t) # should output c(4/7, 2/3, 1)

}
